% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_evaluate.R, R/osi_helpers.R
\name{cf_ind_importance}
\alias{cf_ind_importance}
\title{Helper function to weight and correct the risk and scores}
\usage{
cf_ind_importance(x)

cf_ind_importance(x)
}
\arguments{
\item{x}{The risk or score value to be weighted}

\item{value}{(numeric) The a OSI index value varying between 0 and 1.}
}
\value{
A transformed variable after applying a inverse weighing function so that lower values will gain more impact when applied in a weighed.mean function. A numeric value.

a weighing factor that can be used in calculated a weighted mean
}
\description{
This function gives a value a given weight depending on distance to target
}
\examples{
cf_ind_importance(x = 0.5)
cf_ind_importance(x = c(0.1,0.5,1.5))

}
