% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{oci_c_magnesium_cz}
\alias{oci_c_magnesium_cz}
\title{Calculate the magnesium availability index in Czech Republic}
\usage{
oci_c_magnesium_cz(
  A_MG_M3,
  B_TEXTURE_HYPRES,
  B_LU = NA_character_,
  unitcheck = TRUE
)
}
\arguments{
\item{A_MG_M3}{(numeric) The exchangeable Mg-content of the soil measured via Mehlich 3 extracton (mg Mg/ kg)}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{B_LU}{(character) The crop code}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The magnesium availability index in Czech Republic estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
oci_c_magnesium_cz(B_LU = '3301000000',A_MG_M3 = 81,B_TEXTURE_HYPRES='C')

}
