% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{oci_c_magnesium_ee}
\alias{oci_c_magnesium_ee}
\title{Calculate the magnesium availability index in Estonia}
\usage{
oci_c_magnesium_ee(A_MG_M3, B_TEXTURE_USDA, B_LU = NA_character_)
}
\arguments{
\item{A_MG_M3}{(numeric) The exchangeable Mg-content of the soil measured via Mehlich 3 extracton (mg Mg/ kg)}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{B_LU}{(character) The crop code}
}
\value{
The magnesium availability index in Estonia estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
oci_c_magnesium_ee(A_MG_M3 = 45,B_TEXTURE_USDA = 'clay')

}
