% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_pmn.R
\name{osi_b_pmn_eu}
\alias{osi_b_pmn_eu}
\title{Calculate the index for the microbial biological activity across the EU}
\usage{
osi_b_pmn_eu(B_LU, A_N_RT, A_CLAY_MI)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_N_RT}{(numeric) The total N content of the soil(mg N / kg soil)}

\item{A_CLAY_MI}{(numeric) The aclay content of the soil (\%)}
}
\value{
the normalized potentially mineralizable Nitrogen pool (mg N / kg), a numeric value, converted to an OSI score.
}
\description{
This function assesses the microbial biological activity (of microbes and fungi) via the Potentially Mineralizable N pool, also called PMN (or SoilLife by Eurofins in the past).
}
\examples{
osi_b_pmn_eu(B_LU = 256, A_CLAY_MI = 4.5, A_N_RT = 1250)
osi_b_pmn_eu(c(256,1027),c(4,48), c(3125,1450))

}
