% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_pmn.R
\name{osi_b_pmn_nl}
\alias{osi_b_pmn_nl}
\title{Calculate the index for the microbial biological activity in the Netherlands}
\usage{
osi_b_pmn_nl(B_LU, B_SOILTYPE_AGR, A_N_PMN, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_N_PMN}{(numeric) The potentially mineralizable N pool (mg N / kg soil)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
the normalized potentially mineralizable Nitrogen pool (mg N / kg), a numeric value, converted to an OSI score.
}
\description{
This function assesses the microbial biological activity (of microbes and fungi) via the Potentially Mineralizable N pool, also called PMN (or SoilLife by Eurofins in the past).
}
\examples{
osi_b_pmn_nl(B_LU = '256', B_SOILTYPE_AGR = 'dekzand', A_N_PMN = 125)
osi_b_pmn_nl(c('256','1027'), c('dekzand','rivierklei'), c(125,45))

}
