% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_biodiversity.R
\name{osi_biodiversity}
\alias{osi_biodiversity}
\title{Calculate a biodiversity index}
\usage{
osi_biodiversity(A_SOM_LOI, A_PH_CC)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_PH_CC}{(numeric) The pH measured in CaCl2 solution}
}
\value{
The biodiversity index. A numeric value.
}
\description{
This function calculates the biodiversity index bsed on soil organic matter and pH.
}
\examples{
osi_biodiversity(A_SOM_LOI = 3,A_PH_CC = 4.5)

}
