% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_boron.R
\name{osi_c_boron}
\alias{osi_c_boron}
\title{Calculate the boron availability index (wrapper function)}
\usage{
osi_c_boron(B_LU, A_CLAY_MI, A_B_HW, B_COUNTRY)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_CLAY_MI}{(numeric) The clay content (\%)}

\item{A_B_HW}{(numeric) The plant available content of B in the soil (mg  Cu per kg) extracted by hot water}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The capacity of the soil to supply and buffer boron, evaluated given an optimum threshold for yield. A numeric value.
}
\description{
This function calculates the boron availability for all European countries (if available).
}
