% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_boron.R
\name{osi_c_boron_de}
\alias{osi_c_boron_de}
\title{Calculate the boron availability index in Germany}
\usage{
osi_c_boron_de(B_LU, A_C_OF, A_CLAY_MI, A_SAND_MI, A_PH_CC, A_B_HW)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_C_OF}{(numeric) The carbon content of the soil layer (g/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{A_B_HW}{(numeric) The plant available content of B in the soil (mg B per kg) extracted by hot water}
}
\value{
The boron availability index in Germany derived from extractable soil B fractions. A numeric value.
}
\description{
This function calculates the boron availability.
}
\examples{
osi_c_boron_de(B_LU = 265,A_C_OF=25, A_CLAY_MI=5,A_SAND_MI=15,A_PH_CC = 4,A_B_HW = 50)
 
}
