% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_boron.R
\name{osi_c_boron_pt}
\alias{osi_c_boron_pt}
\title{Calculate the boron availability index in Portugal}
\usage{
osi_c_boron_pt(B_LU, A_B_HW, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_B_HW}{(numeric) The plant available content of B in the soil (mg B per kg) extracted by hot water}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The boron availability index in Portugal derived from extractable soil B fractions. A numeric value.
}
\description{
This function calculates the boron availability.
}
\examples{
osi_c_boron_pt(B_LU = '3301010901',A_B_HW = 50)
osi_c_boron_pt(B_LU = c('3301010901','3304990000'),A_B_HW = c(35,55))

}
