% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_boron.R
\name{osi_c_boron_se}
\alias{osi_c_boron_se}
\title{Calculate the boron availability index in Sweden}
\usage{
osi_c_boron_se(B_LU, A_PH_WA)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_PH_WA}{(numeric) The pH measured in water}
}
\value{
The boron availability index in Sweden depends primarily on pH. A numeric value.
}
\description{
This function calculates the boron availability.
}
\examples{
osi_c_boron_ie(B_LU = 265,A_PH_WA = 5.0)
osi_c_boron_ie(B_LU = c(265,1019),A_PH_WA = c(3.5,5.5))

}
