% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_boron.R
\name{osi_c_boron_uk}
\alias{osi_c_boron_uk}
\title{Calculate the boron availability index in United Kingdom}
\usage{
osi_c_boron_uk(B_LU, B_TEXTURE_HYPRES, A_SOM_LOI, A_PH_CC, A_B_HW)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{A_B_HW}{(numeric) The plant available content of B in the soil (mg B per kg) extracted by hot water}
}
\value{
The boron availability index in UK derived from extractable soil B fractions. A numeric value.
}
\description{
This function calculates the boron availability.
}
\examples{
osi_c_boron_uk(B_LU = '265',B_TEXTURE_HYPRES='C',A_SOM_LOI=3,
A_PH_CC = 4,A_B_HW = 50)
osi_c_boron_uk(B_LU = c('265','1019'),B_TEXTURE_HYPRES = c('C','F'),
A_SOM_LOI = c(3,3),A_PH_CC = c(4,6),A_B_HW = c(35,55))

}
