% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_copper.R
\name{osi_c_copper_ro}
\alias{osi_c_copper_ro}
\title{Calculate the Cu availability index for agricultural soils in Romania}
\usage{
osi_c_copper_ro(B_LU, A_CU_EDTA)
}
\arguments{
\item{B_LU}{(character) crop type}

\item{A_CU_EDTA}{(numeric) Cu content measured in EDTA (mg / kg)}
}
\value{
The copper availability index in Romania estimated from extractable Cu with 0.05M Na2EDTA, a numeric value.
}
\description{
This function calculates the Cu availability of a soil, using the agronomic index used in Romania
}
\examples{
osi_c_copper_ro(B_LU = 'testcrop', A_CU_EDTA = 45)

}
