% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium}
\alias{osi_c_magnesium}
\title{Calculate the magnesium availability index (wrapper function)}
\usage{
osi_c_magnesium(
  B_LU,
  B_SOILTYPE_AGR = NA_character_,
  A_SOM_LOI = NA_real_,
  A_CLAY_MI = NA_real_,
  A_PH_CC = NA_real_,
  A_CACO3_IF = NA_real_,
  A_CEC_CO = NA_real_,
  A_K_CO_PO = NA_real_,
  A_K_CC = NA_real_,
  A_MG_AA = NA_real_,
  A_MG_CC = NA_real_,
  B_COUNTRY
)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) is the Clay content (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{A_CACO3_IF}{(numeric) the percentage of CaCO3 (\%)}

\item{A_CEC_CO}{(numeric) is the Cation exhange capacity in mmol+/kg}

\item{A_K_CO_PO}{(numeric) The occupation of the CEC with potassium (\%)}

\item{A_K_CC}{(numeric) The plant available potassium, extracted with 0.01M CaCl2 (mg per kg)}

\item{A_MG_AA}{(numeric) is the exchangeable Mg concentration (mg/kg)}

\item{A_MG_CC}{(numeric) The plant available content of Mg in the soil (mg  Mg per kg) extracted by 0.01M CaCl2}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The capacity of the soil to supply and buffer potassium, evaluated given an optimum threshold for yield. A numeric value.
}
\description{
This function calculates the magnesium availability for all European countries (if available).
}
