% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_ch}
\alias{osi_c_magnesium_ch}
\title{Calculate the magnesium availability index in Switzerland}
\usage{
osi_c_magnesium_ch(A_MG_AAA, A_CLAY_MI, B_LU = NA_character_)
}
\arguments{
\item{A_MG_AAA}{(numeric) The exchangeable Mg-content of the soil measured via acid ammonium acetate extraction (mg Mg/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium availability index in Switzerland estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_ch(A_MG_AAA = 50)

}
