% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_de}
\alias{osi_c_magnesium_de}
\title{Calculate the magnesium availability index in Germany}
\usage{
osi_c_magnesium_de(B_LU, A_C_OF, A_CLAY_MI, A_SAND_MI, A_MG_CC)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_C_OF}{(numeric) The carbon content of the soil layer (g/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_MG_CC}{(numeric) The magnesium content extracted with CaCl2 (g / kg)}
}
\value{
The magnesium availability index in Germany derived from extractable soil Mg fractions. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
