% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_dk}
\alias{osi_c_magnesium_dk}
\title{Calculate the magnesium availability index in Denmark}
\usage{
osi_c_magnesium_dk(B_LU, A_MG_AL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_MG_AL}{(numeric) The Mg-content of the soil extracted with ammonium lactate (mg Mg / kg)}
}
\value{
The magnesium availability index in Denmark derived from extractable soil Mg fractions. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_dk(B_LU = '3301000000',A_MG_AL = 5)
osi_c_magnesium_dk(B_LU = c('3301000000','3301061299'),A_MG_AL = c(3.5,5.5))

}
