% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_fi}
\alias{osi_c_magnesium_fi}
\title{Calculate the magnesium availability index in Finland}
\usage{
osi_c_magnesium_fi(
  B_LU,
  B_TEXTURE_USDA,
  A_MG_AAA,
  A_C_OF = 0.5,
  unitcheck = TRUE
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{A_MG_AAA}{(numeric) The exchangeable Mg-content of the soil measured via ammonium acetate extraction (mg Mg / kg)}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The magnesium availability index in Finland estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_fi(B_LU = '1110', B_TEXTURE_USDA = 'Si',A_MG_AAA = 45)

}
