% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_fr}
\alias{osi_c_magnesium_fr}
\title{Calculate the Mg availability index for France}
\usage{
osi_c_magnesium_fr(B_LU, A_CLAY_MI, A_CEC_CO, A_CACO3_IF, A_MG_AA)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_CLAY_MI}{(numeric) Soil clay content (\%)}

\item{A_CEC_CO}{(numeric) The cation exchange capacity of the soil in mmol/kg}

\item{A_CACO3_IF}{(numeric) The CaCO3 content in the soil (\%)}

\item{A_MG_AA}{(numeric) The extractable Mg content in the soil (mg/kg)}
}
\value{
The magnesium availability index in France estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the Mg availability of a soil, using the agronomic index used in France
}
