% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_hu}
\alias{osi_c_magnesium_hu}
\title{Calculate the magnesium availability index in Hungary}
\usage{
osi_c_magnesium_hu(
  B_TEXTURE_USDA,
  A_MG_KCL,
  B_LU = NA_character_,
  unitcheck = TRUE
)
}
\arguments{
\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{A_MG_KCL}{(numeric) The exchangeable Mg-content of the soil measured via KCL extracton (mg Mg/ kg)}

\item{B_LU}{(character) The crop code}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The magnesium availability index in Hungary estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_hu(B_LU = 'testcrop',A_MG_KCL = 45,B_TEXTURE_USDA = 'loam')

}
