% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_ie}
\alias{osi_c_magnesium_ie}
\title{Calculate the magnesium availability index in Ireland}
\usage{
osi_c_magnesium_ie(B_LU, A_SOM_LOI, A_MG_NaAAA)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_MG_NaAAA}{(numeric) The Mg-content of the soil extracted with Morgan's solution, sodium acetate acetic acid (mg/ kg)}
}
\value{
The magnesium availability index in Ireland derived from extractable soil Mg fractions. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_ie(B_LU = 'testcrop',A_SOM_LOI = 2,A_MG_NaAAA = 5)
osi_c_magnesium_ie(B_LU = c('testcrop','testcrop2'),A_SOM_LOI = c(2,4),A_MG_NaAAA = c(3.5,5.5))

}
