% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_lt}
\alias{osi_c_magnesium_lt}
\title{Calculate the magnesium availability index in Lithuania}
\usage{
osi_c_magnesium_lt(A_MG_AL, A_PH_KCL, B_LU = NA_character_)
}
\arguments{
\item{A_MG_AL}{(numeric) The exchangeable Mg-content of the soil measured via Ammonium Lactate extraction (mg Mg/ kg)}

\item{A_PH_KCL}{(numeric) The pH measured in KCl}

\item{B_LU}{(character) The crop code}
}
\value{
The magnesium availability index in Lithuania estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_lt(A_MG_AL = 45,A_PH_KCL= 5.5)

}
