% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_lv}
\alias{osi_c_magnesium_lv}
\title{Calculate the magnesium availability index in Latvia}
\usage{
osi_c_magnesium_lv(A_MG_DL, B_TEXTURE_USDA, B_LU = NA_character_)
}
\arguments{
\item{A_MG_DL}{(numeric) The exchangeable Mg-content of the soil measured via Double Lactate extraction (mg Mg/ kg)}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{B_LU}{(character) The crop code}
}
\value{
The magnesium availability index in Latvia estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_lv(A_MG_DL = 45,B_TEXTURE_USDA='sand')

}
