% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_ro}
\alias{osi_c_magnesium_ro}
\title{Calculate the magnesium availability index in Romenia}
\usage{
osi_c_magnesium_ro(
  B_LU,
  B_TEXTURE_HYPRES = B_TEXTURE_HYPRES,
  A_CEC_CO,
  A_MG_CC,
  A_MG_CO_PO,
  A_K_CO_PO,
  A_PH_WA
)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{A_CEC_CO}{(numeric) is the Cation exhange capacity in mmol+/kg}

\item{A_MG_CC}{(numeric) The plant available content of Mg in the soil (mg  Mg per kg) extracted by 0.01M CaCl2}

\item{A_MG_CO_PO}{(numeric) The exchangeable Mg-content of the soil measured via Cohex extracton, percentage occupation at CEC (\%)}

\item{A_K_CO_PO}{(numeric) The exchangeable K-content of the soil measured via Cohex extracton, percentage occupation at CEC (\%)}

\item{A_PH_WA}{(numeric) The pH measured in water}
}
\value{
The Mg index in Romenia. A numeric value.
}
\description{
This function calculates the magnesium availability index
}
\examples{
osi_c_magnesium_ro(B_LU = 'testcrop1',B_TEXTURE_HYPRES ='M', A_CEC_CO = 140,
A_MG_CC = 60, A_MG_CO_PO = 8,A_K_CO_PO = 12, A_PH_WA = 5)

}
