% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_se}
\alias{osi_c_magnesium_se}
\title{Calculate the magnesium availability index in Sweden}
\usage{
osi_c_magnesium_se(A_MG_AL, B_LU = NA_character_)
}
\arguments{
\item{A_MG_AL}{(numeric) The exchangeable Mg-content of the soil measured via ammonium lactate (mg K/kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The magnesium availability index in Sweden estimated from the Mg extracted soil pool. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_se(A_MG_AL = 45)

}
