% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_sk}
\alias{osi_c_magnesium_sk}
\title{Calculate the magnesium availability index in  Slovak Republic}
\usage{
osi_c_magnesium_sk(B_TEXTURE_HYPRES, A_MG_M3, B_LU = NA_character_)
}
\arguments{
\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{B_LU}{(character) The crop code}

\item{A_K_M3}{(numeric) The exchangeable Mg-content of the soil measured via Mehlich 3 extracton (mg Mg/ kg)}
}
\value{
The magnesium availability index in Slovak Republic estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_sk(A_MG_M3 = 45,B_TEXTURE_HYPRES='C')

}
