% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_sl}
\alias{osi_c_magnesium_sl}
\title{Calculate the magnesium availability index in Slovenia}
\usage{
osi_c_magnesium_sl(
  A_MG_AL,
  B_TEXTURE_HYPRES,
  B_LU = NA_character_,
  unitcheck = TRUE
)
}
\arguments{
\item{A_MG_AL}{(numeric) The exchangeable Mg-content of the soil measured via ammonium lactate extracton (mg Mg/ kg)}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{B_LU}{(character) The crop code}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The magnesium availability index in Slovenia estimated from extractable magnesium. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_sl(A_MG_AL = 45,B_TEXTURE_HYPRES='C')

}
