% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_uk}
\alias{osi_c_magnesium_uk}
\title{Calculate the magnesium availability index in United Kingdom}
\usage{
osi_c_magnesium_uk(B_LU, A_SOM_LOI, A_MG_AN)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_MG_AN}{(numeric) The Mg-content of the soil extracted with ammonium nitrate (mg Mg /kg)}
}
\value{
The magnesium availability index in United Kingdom derived from extractable soil Mg fractions. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_uk(B_LU = 265,A_SOM_LOI=3,A_MG_AN = 50)
osi_c_magnesium_uk(B_LU = c(265,1019),A_SOM_LOI = c(3,5),A_MG_AN = c(35,55))

}
