% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen.R
\name{osi_c_nitrogen}
\alias{osi_c_nitrogen}
\title{Calculate the soil nitrogen supplying capacity (wrapper function)}
\usage{
osi_c_nitrogen(B_LU, B_SOILTYPE_AGR, A_SOM_LOI, A_N_RT, B_COUNTRY)
}
\arguments{
\item{B_LU}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{B_COUNTRY}{(character) The country code}
}
\value{
the normalized potentially mineralizable Nitrogen pool (mg N / kg), a numeric value, converted to an OSI score.
}
\description{
This function calculates the nitrogen supplying capacity for soils in Europe
}
\examples{
osi_c_nitrogen(B_LU = 256, B_SOILTYPE_AGR = 'dekzand',A_N_RT = 2500, 
A_SOM_LOI = 4.5, B_COUNTRY = 'NL')
osi_c_nitrogen(1019,'dekzand',4.5,2315,'NL')

}
