% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen.R
\name{osi_c_nitrogen_be}
\alias{osi_c_nitrogen_be}
\title{Calculate the soil nitrogen supplying capacity in Belgium}
\usage{
osi_c_nitrogen_be(B_LU, A_N_RT, A_C_OF, A_CLAY_MI, A_SAND_MI, A_CACO3_IF)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_CACO3_IF}{(numeric) The percentage of carbonated lime (\%)}
}
\value{
The capacity of the soil to supply nitrogen (kg N / ha / yr). A numeric value, converted to a OSI score.
}
\description{
This function calculates the NSC (nitrogen supply capacity) for the soil
}
\examples{
osi_c_nitrogen_be(B_LU = '772',A_N_RT = 1200, A_C_OF = 25, 
A_CLAY_MI = 3.5, A_SAND_MI = 15,A_CACO3_IF = 0.8)

}
