% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen.R
\name{osi_c_nitrogen_de}
\alias{osi_c_nitrogen_de}
\title{Calculate the soil nitrogen supplying capacity in Germany}
\usage{
osi_c_nitrogen_de(
  B_LU = NA_character_,
  A_CLAY_MI = NA_real_,
  A_SAND_MI = NA_real_,
  A_SOM_LOI = NA_real_,
  A_C_OF = NA_real_,
  A_N_RT = NA_real_
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}
}
\value{
The capacity of the soil to supply nitrogen (kg N / ha / yr). A numeric value, converted to a OSI score.
}
\description{
This function calculates the nitrogen producing capacity of the soil, and applies correction factors used affecting N availability.
}
\examples{
osi_c_nitrogen_de(B_LU= '3301010100', A_CLAY_MI = 25, 
A_SAND_MI = 7.5, A_SOM_LOI = 4.5, A_N_RT = 2500)

}
