% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen.R
\name{osi_c_nitrogen_nl}
\alias{osi_c_nitrogen_nl}
\title{Calculate the soil nitrogen supplying capacity in the Netherlands}
\usage{
osi_c_nitrogen_nl(B_LU, B_SOILTYPE_AGR, A_SOM_LOI, A_N_RT, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The capacity of the soil to supply nitrogen (kg N / ha / yr). A numeric value, converted to a OSI score.
}
\description{
This function calculates the NLV (nitrogen producing capacity) for the soil
}
\examples{
osi_c_nitrogen_nl(B_LU = '256', B_SOILTYPE_AGR = 'dekzand',
A_SOM_LOI = 4.5,A_N_RT = 2500)
osi_c_nitrogen_nl(1019,'dekzand',5.5,2315)

}
