% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph}
\alias{osi_c_ph}
\title{Calculate the pH index (wrapper function)}
\usage{
osi_c_ph(A_PH_WA, A_PH_CC, A_PH_KCL, B_LU, B_TEXTURE_USDA)
}
\arguments{
\item{A_PH_WA}{(numeric) The pH measured in h2o}

\item{A_PH_CC}{(numeric) The pH measured in cacl2}

\item{A_PH_KCL}{(numeric) The pH measured in KCl}

\item{B_LU}{(numeric) The crop code}

\item{B_TEXTURE_USDA}{(character) The USDA textural class}
}
\value{
The index to evaluate the soil pH
}
\description{
This function calculates the pH index for all European countries (if available).
}
