% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_at}
\alias{osi_c_ph_at}
\title{Calculate the pH index in Austria}
\usage{
osi_c_ph_at(B_LU, A_PH_CC, B_TEXTURE_HYPRES, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_PH_CC}{(numeric) The pH measured in cacl2}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The pH index in Austria, depending on land use and soil type. A numeric value.
}
\description{
This function calculates the soil acidity
}
\examples{
osi_c_ph_at(B_LU = '3301000000', A_PH_CC = 4.7,B_TEXTURE_HYPRES = 'C')

}
