% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_be}
\alias{osi_c_ph_be}
\title{Calculate the pH index for Belgium}
\usage{
osi_c_ph_be(B_LU, B_TEXTURE_BE, A_PH_KCL)
}
\arguments{
\item{B_LU}{(character) The crop type}

\item{B_TEXTURE_BE}{(character) The soil texture according to Belgium classification system}

\item{A_PH_KCL}{(numeric) The pH measured in KCl}
}
\value{
The pH index in Belgium estimated from pH in KCL, the textural class and the crop type
}
\description{
This function evaluates the pH index in Belgium
}
