% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_ch}
\alias{osi_c_ph_ch}
\title{Calculate the pH index for Switzerland}
\usage{
osi_c_ph_ch(
  B_LU,
  A_CLAY_MI = NA_real_,
  A_PH_WA = NA_real_,
  A_CA_CO_PO = NA_real_,
  A_MG_CO_PO = NA_real_,
  A_K_CO_PO = NA_real_,
  A_NA_CO_PO = NA_real_
)
}
\arguments{
\item{B_LU}{(character) The crop type}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_WA}{(numeric) The pH measured in h2o}

\item{A_CA_CO_PO}{(numeric) The calcium occupation of the CEC (\%)}

\item{A_MG_CO_PO}{(numeric) The magnesium occupation of the CEC (\%)}

\item{A_K_CO_PO}{(numeric) The potassium occupation of the CEC (\%)}

\item{A_NA_CO_PO}{(numeric) The sodium occupation of the CEC (\%)}
}
\value{
The pH index in Switzerland estimated from the base saturation with Ca, Mg and K. For soils with pH water below 5.9 the soil is extracted with HCL+H2SO4. Soils with higher pH are extracted with Bariumchloride.
}
\description{
This function evaluates the pH index in Switzerland
}
