% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_de}
\alias{osi_c_ph_de}
\title{Calculate the pH index for soils in Germany}
\usage{
osi_c_ph_de(B_LU, A_SOM_LOI, A_C_OF, A_CLAY_MI, A_SAND_MI, A_PH_CC)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_C_OF}{(numeric) The carbon content of the soil layer (g/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_PH_CC}{(numeric) The soil acidity measures with CaCl2 method.}
}
\value{
The pH index in Germany derived for grassland and arable crops. A numeric value.
}
\description{
This function calculates the soil pH.
}
