% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_el}
\alias{osi_c_ph_el}
\title{Calculate the pH index in Greece}
\usage{
osi_c_ph_el(B_LU, A_PH_WA, A_NA_CO_PO, B_TEXTURE_HYPRES, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_PH_WA}{(numeric) The pH measured in water}

\item{A_NA_CO_PO}{(numeric) The sodium occupation of the CEC (\%)}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The pH index in Greece. A numeric value.
}
\description{
This function calculates the pH index
}
\examples{
osi_c_ph_el(B_LU = 'testcrop1',A_PH_WA = 5,A_NA_CO_PO = 3,B_TEXTURE_HYPRES='C')
osi_c_ph_el(B_LU = c('testcrop1','testcrop2'),A_PH_WA = c(3.5,5.5),
A_NA_CO_PO = c(1,6),B_TEXTURE_HYPRES=c('C','M'))

}
