% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_fi}
\alias{osi_c_ph_fi}
\title{Calculate the pH index for Finland}
\usage{
osi_c_ph_fi(B_LU, B_TEXTURE_USDA, A_PH_WA, A_C_OF = 0.5, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{A_PH_WA}{(numeric) The pH values determined via water extract}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The pH index in Finland estimated from pH water. A numeric value.
}
\description{
This function evaluates the pH index in Finland
}
\examples{
osi_c_ph_fi(B_LU = '1110', B_TEXTURE_USDA = 'Si',A_PH_WA = 4.5)

}
