% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_ro}
\alias{osi_c_ph_ro}
\title{Calculate the pH index in Romenia}
\usage{
osi_c_ph_ro(B_LU, A_PH_WA, A_NA_CO_PO)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_PH_WA}{(numeric) The pH measured in water}

\item{A_NA_CO_PO}{(numeric) The sodium occupation of the CEC (\%)}
}
\value{
The pH index in Romenia. A numeric value.
}
\description{
This function calculates the pH index
}
\examples{
osi_c_ph_ro(B_LU = 'testcrop1',A_PH_WA = 5,A_NA_CO_PO=3)
osi_c_ph_ro(B_LU = c('testcrop1','testcrop2'),A_PH_WA = c(3.5,5.5),A_NA_CO_PO=c(1,3))

}
