% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_se}
\alias{osi_c_ph_se}
\title{Calculate the pH index in Sweden}
\usage{
osi_c_ph_se(B_LU, A_SOM_LOI, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The organic matter content of soil in percentage}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_WA}{(numeric) The pH measured in water}
}
\value{
The pH index in Sweden. A numeric value.
}
\description{
This function calculates the pH index
}
\examples{
osi_c_ph_se(B_LU = '3301010901',A_PH_WA = 5,A_CLAY_MI = 15, A_SOM_LOI = 2)
osi_c_ph_se(B_LU = c('3301010901','3304990000'),A_PH_WA = c(3.5,5.5),
A_CLAY_MI = c(5,15),A_SOM_LOI = c(2,4))

}
