% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_uk}
\alias{osi_c_ph_uk}
\title{Calculate the pH index in United Kingdom}
\usage{
osi_c_ph_uk(B_LU, A_PH_WA, A_SOM_LOI, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_PH_WA}{(numeric) The pH measured in water}

\item{A_SOM_LOI}{(numeric) The organic matter content of soil in percentage}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The pH index in UK A numeric value.
}
\description{
This function calculates the pH index
}
\examples{
osi_c_ph_uk(B_LU = 'testcrop1',A_PH_WA = 5, A_SOM_LOI = 4)
osi_c_ph_uk(B_LU = c('testcrop1','testcrop2'),A_PH_WA = c(3.5,5.5), A_SOM_LOI = c(3.5,4))

}
