% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_at}
\alias{osi_c_phosphor_at}
\title{Calculate the phosphorus availability index in Austria}
\usage{
osi_c_phosphor_at(B_LU, A_P_CAL, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_P_CAL}{(numeric) The exchangeable P-content of the soil measured via Calcium Ammonium Lactate (mg P/ kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphorus availability index in Austria estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus availability.
}
\examples{
osi_c_phosphor_at(B_LU = '3301000000', A_P_CAL = 47)

}
