% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_be}
\alias{osi_c_phosphor_be}
\title{Calculate the phosphorus availability index in Belgium}
\usage{
osi_c_phosphor_be(B_LU, A_P_AL, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_P_AL}{(numeric) The exchangeable P-content of the soil measured via ammonium lactate extraction (mg P/ kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphorus availability index in Belgium estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus availability.
}
\examples{
osi_c_phosphor_be(B_LU = '8410', A_P_AL = 45)

}
