% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_cz}
\alias{osi_c_phosphor_cz}
\title{Calculate the phosphorus availability index in Czech Republic}
\usage{
osi_c_phosphor_cz(B_LU, A_P_M3)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_P_M3}{(numeric) The exchangeable P-content of the soil measured via Mehlich 3 extracton (mg P/ kg)}
}
\value{
The phosphorus availability index in Czech Republic estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus availability.
}
\examples{
osi_c_phosphor_cz(B_LU = '3303030100', A_P_M3 = 81)

}
