% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_el}
\alias{osi_c_phosphor_el}
\title{Calculate the phosphate availability index in Greece}
\usage{
osi_c_phosphor_el(B_LU, A_P_OL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg P/kg)}
}
\value{
The phosphate availability index in Greece derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate availability.
}
\examples{
osi_c_phosphor_el(B_LU = '3301061299',A_P_OL = 5)
osi_c_phosphor_el(B_LU = c('3301061299','3301000000'),A_P_OL = c(5,15))

}
