% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_fi}
\alias{osi_c_phosphor_fi}
\title{Calculate the phosphorus availability index in Finland}
\usage{
osi_c_phosphor_fi(
  B_LU,
  B_TEXTURE_USDA,
  A_P_AAA,
  A_C_OF = 0.5,
  unitcheck = TRUE
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{A_P_AAA}{(numeric) The exchangeable P-content of the soil measured via acid ammonium acetate extraction}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphorus availability index in Finland estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus availability.
}
\examples{
osi_c_phosphor_fi(B_LU = '4010', B_TEXTURE_USDA = 'Si',
A_P_AAA = 45,A_C_OF=1.5)

}
