% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_fr}
\alias{osi_c_phosphor_fr}
\title{Calculate the phosphate availability index in France}
\usage{
osi_c_phosphor_fr(B_LU, A_P_OL, A_PH_WA = NA_real_, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg P/ kg)}

\item{A_PH_WA}{(numeric) The pH measured in water.}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphate availability index in France estimated from extractable soil P Olsen (a numeric value).
}
\description{
This function calculates the phosphate availability.
}
\examples{
osi_c_phosphor_fr(B_LU = 'SOJ', A_P_OL = 45)

}
