% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_uk}
\alias{osi_c_phosphor_uk}
\title{Calculate the phosphate availability index in United Kingdom}
\usage{
osi_c_phosphor_uk(B_LU, A_SOM_LOI, A_P_OL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg P /kg)}
}
\value{
The phosphate availability index in United Kingdom derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate availability.
}
\examples{
osi_c_phosphor_uk(B_LU = 265,A_SOM_LOI = 3,A_P_OL = 5)
osi_c_phosphor_uk(B_LU = c(265,1019),A_SOM_LOI = c(3,4),A_P_OL = c(3.5,5.5))

}
