% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_posphor_at}
\alias{osi_c_posphor_at}
\title{Calculate the phosphorus availability index in Austria}
\usage{
osi_c_posphor_at(A_P_CAL, B_LU = NA_character_)
}
\arguments{
\item{A_P_CAL}{(numeric) The exchangeable P-content of the soil measured via Calcium Ammonium Lactate (mg P/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus availability index in Austria estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus availability.
}
\examples{
osi_c_phosphorus_at(A_P_CAL = 47)

}
