% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_posphor_fr}
\alias{osi_c_posphor_fr}
\title{Calculate the phosphate availability index in France}
\usage{
osi_c_posphor_fr(B_LU, B_SOILTYPE_AGR, B_AER_FR, A_P_OL = NA_real_)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_SOILTYPE_AGR}{(character) The soil type in a particular region}

\item{B_AER_FR}{(character) An agroeconomic region in France}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen}
}
\value{
The phosphate availability index in France estimated from extractable soil P Olsen (a numeric value).
}
\description{
This function calculates the phosphate availability.
}
\examples{
osi_c_posphor_fr(B_LU = 'SOJ', A_P_OL = 45, 
B_SOILTYPE_AGR = 'limons battants', B_AER_FR = 'nord-picardie')

}
