% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium}
\alias{osi_c_potassium}
\title{Calculate the K availability index (wrapper function)}
\usage{
osi_c_potassium(
  B_LU,
  B_SOILTYPE_AGR = NA_character_,
  B_AER_FR = NA_character_,
  A_SOM_LOI = NA,
  A_CLAY_MI = NA,
  A_PH_CC = NA,
  A_CEC_CO = NA,
  A_K_CO_PO = NA,
  A_K_CC = NA,
  A_K_AA = NA,
  B_COUNTRY
)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_AER_FR}{(character) An agroeconomic region in France}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{A_CEC_CO}{(numeric) The cation exchange capacity of the soil (mmol+ / kg), analyzed via Cobalt-hexamine extraction}

\item{A_K_CO_PO}{(numeric) The occupation of the CEC with potassium (\%)}

\item{A_K_CC}{(numeric) The plant available potassium, extracted with 0.01M CaCl2 (mg / kg),}

\item{A_K_AA}{(numeric) The exchangeable K-content of the soil measured via ammonium acetate extraction}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The capacity of the soil to supply and buffer potassium, evaluated given an optimum threshold for yield. A numeric value.
}
\description{
This function calculates the potassium availability for all European countries (if available).
}
\examples{
osi_c_potassium(B_LU = 265, B_SOILTYPE_AGR = 'dekzand',A_SOM_LOI = 4, 
A_CLAY_MI = 11,A_PH_CC = 5.4, A_CEC_CO = 125, 
A_K_CO_PO = 8.5, A_K_CC = 145,B_COUNTRY = 'NL')

}
