% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_ch}
\alias{osi_c_potassium_ch}
\title{Calculate the potassium availability index in Switzerland}
\usage{
osi_c_potassium_ch(A_K_AAA, A_CLAY_MI, B_LU = NA_character_)
}
\arguments{
\item{A_K_AAA}{(numeric) The exchangeable K-content of the soil measured via acid ammonium acetate extraction (mg K/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium availability index in Switzerland estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_ch(A_K_AAA = 50, A_CLAY_MI = 25)

}
