% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_ee}
\alias{osi_c_potassium_ee}
\title{Calculate the potassium availability index in Estonia}
\usage{
osi_c_potassium_ee(A_K_M3, B_TEXTURE_USDA, B_LU = NA_character_)
}
\arguments{
\item{A_K_M3}{(numeric) The exchangeable K-content of the soil measured via Mehlich 3 extracton (mg P/ kg)}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium availability index in Estonia estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_ee(B_LU = 'testcrop',A_K_M3 = 45,B_TEXTURE_USDA = 'ClLo')

}
