% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_fi}
\alias{osi_c_potassium_fi}
\title{Calculate the potassium availability index in Finland}
\usage{
osi_c_potassium_fi(B_LU, B_TEXTURE_USDA, A_K_AAA, A_C_OF = 0)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{A_K_AAA}{(numeric) The exchangeable K-content of the soil measured via ammonium acetate extraction}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}
}
\value{
The potassium availability index in Finland estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_fi(B_LU = 'SOJ', B_TEXTURE_USDA = 'Si',A_K_AAA = 45,A_C_OF = 15)

}
