% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_fr}
\alias{osi_c_potassium_fr}
\title{Calculate the potassium availability index in France}
\usage{
osi_c_potassium_fr(B_LU, B_SOILTYPE_AGR, B_AER_FR, A_K_AA = NA_real_)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_SOILTYPE_AGR}{(character) The soil type in a particular region}

\item{B_AER_FR}{(character) An agroeconomic region in France}

\item{A_K_AA}{(numeric) The exchangeable K-content of the soil measured via ammonium acetate extraction}
}
\value{
The potassium availability index in France estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_fr(B_LU = 'SOJ', A_K_AA = 45,
B_SOILTYPE_AGR = 'limons battants', B_AER_FR = 'nord picardie')

}
