% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_lt}
\alias{osi_c_potassium_lt}
\title{Calculate the potassium availability index in Lithuania}
\usage{
osi_c_potassium_lt(A_K_AL, A_SOM_LOI, B_LU = NA_character_, unitcheck = TRUE)
}
\arguments{
\item{A_K_AL}{(numeric) The exchangeable K-content of the soil measured via Ammonium Lactate extraction (mg P/ kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{B_LU}{(character) The crop code}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium availability index in Lithuania estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_lt(A_K_AL = 45,A_SOM_LOI= 1.5)

}
