% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_no}
\alias{osi_c_potassium_no}
\title{Calculate the potassium availability index in Norway}
\usage{
osi_c_potassium_no(B_LU, A_K_AL, A_CLAY_MI, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_K_AL}{(numeric) The K-content of the soil extracted with ammonium lactate (mg K / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium availability index in Norway derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_no(B_LU = 'testcrop1',A_K_AL = 5,A_CLAY_MI=5)
osi_c_potassium_no(B_LU = c('testcrop1','testcrop2'),A_K_AL = c(3.5,5.5),A_CLAY_MI=c(3,5))

}
