% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_pl}
\alias{osi_c_potassium_pl}
\title{Calculate the potassium availability index in Poland}
\usage{
osi_c_potassium_pl(A_K_DL, B_TEXTURE_HYPRES, B_LU = NA_character_)
}
\arguments{
\item{A_K_DL}{(numeric) The exchangeable K-content of the soil measured via ammonium double lactate extracton (mg K/ kg)}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium availability index in Poland estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_pl(B_LU = 'testcrop1',A_K_DL = 45,B_TEXTURE_HYPRES='C')

}
