% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_se}
\alias{osi_c_potassium_se}
\title{Calculate the potassium availability index in Sweden}
\usage{
osi_c_potassium_se(B_LU, A_K_AL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_K_AL}{(numeric) The K-content of the soil extracted with ammonium lactate (mg K / kg)}
}
\value{
The potassium availability index in Sweden derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_se(B_LU = 265,A_K_AL = 5)
osi_c_potassium_se(B_LU = c(265,1019),A_K_AL = c(3.5,5.5))

}
