% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_zinc.R
\name{osi_c_zinc_fr}
\alias{osi_c_zinc_fr}
\title{Calculate the Zn availability index for agricultural soils in France}
\usage{
osi_c_zinc_fr(B_LU, A_PH_WA, A_ZN_EDTA, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) crop type}

\item{A_PH_WA}{(numeric) pH measured in water (-)}

\item{A_ZN_EDTA}{(numeric) Zn content measured in EDTA (mg / kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The zinc availability index in France estimated from extractable zinc and pH measured in water, a numeric value.
}
\description{
This function calculates the Zn availability of a soil, using the agronomic index used in France.
}
\examples{
osi_c_zinc_fr(B_LU = 'SOJ', A_ZN_EDTA = 45, A_PH_WA = 6.5)

}
