% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_zinc.R
\name{osi_c_zinc_ie}
\alias{osi_c_zinc_ie}
\title{Calculate the Zn availability index for agricultural soils in Ireland}
\usage{
osi_c_zinc_ie(B_LU, A_SOM_LOI, A_PH_WA, A_ZN_EDTA)
}
\arguments{
\item{B_LU}{(character) crop type}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_PH_WA}{(numeric) pH measured in water (-)}

\item{A_ZN_EDTA}{(numeric) Zn content measured in EDTA (mg / kg)}
}
\value{
The zinc availability index in Ireland estimated from extractable zinc and pH measured in water, a numeric value.
}
\description{
This function calculates the Zn availability of a soil, using the agronomic index used in Ireland.
}
\examples{
osi_c_zinc_ie(B_LU = 'SOJ', A_SOM_LOI = 4, A_ZN_EDTA = 45, A_PH_WA = 6.5)

}
