% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_zinc.R
\name{osi_c_zinc_nl}
\alias{osi_c_zinc_nl}
\title{Calculate the Zn availability index for the Netherlands}
\usage{
osi_c_zinc_nl(B_LU, A_PH_CC, A_ZN_CC)
}
\arguments{
\item{B_LU}{(numeric) The crop code from the BRP}

\item{A_PH_CC}{(numeric) The acidity of the soil, determined in 0.01M CaCl2 (-)}

\item{A_ZN_CC}{The plant available Zn content, extracted with 0.01M CaCl2 (mg / kg)}
}
\value{
The function of the soil to supply zinc (a numeric value).
}
\description{
This function calculates the availability of Zn for plant uptake
}
\examples{
osi_c_zinc_nl(B_LU = 265, A_ZN_CC = 45, A_PH_CC = 6.5)

}
