% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_zinc.R
\name{osi_c_zinc_pt}
\alias{osi_c_zinc_pt}
\title{Calculate the Zn availability index for Portugal}
\usage{
osi_c_zinc_pt(B_LU, A_ZN_AAA, A_PH_CC)
}
\arguments{
\item{B_LU}{(numeric) A unique crop code}

\item{A_ZN_AAA}{The plant available Zn content, extracted with ammonium acetate (mg Zn / kg)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}
}
\value{
The function of the soil to supply zinc (a numeric value).
}
\description{
This function calculates the availability of Zn for plant uptake
}
\examples{
osi_c_zinc_pt(B_LU = '3301061299', A_ZN_AAA = 45,A_PH_CC = 4.5)

}
