% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_zinc.R
\name{osi_c_zinc_ro}
\alias{osi_c_zinc_ro}
\title{Calculate the Zn availability index for agricultural soils in Romenia}
\usage{
osi_c_zinc_ro(B_LU, A_PH_WA, A_P_AL, A_ZN_EDTA)
}
\arguments{
\item{B_LU}{(character) crop type}

\item{A_PH_WA}{(numeric) pH measured in water (-)}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with ammonium lactate (mg P / kg)}

\item{A_ZN_EDTA}{(numeric) Zn content measured in EDTA, pH 8.6, with ammonium carbonate (mg / kg)}
}
\value{
The zinc availability index in United Kingdom estimated from extractable zinc and pH measured in water, a numeric value.
}
\description{
This function calculates the Zn availability of a soil, using the agronomic index used in RO.
}
\examples{
osi_c_zinc_ro(B_LU = 'testcrop', A_ZN_EDTA = 45, A_P_AL = 45,A_PH_WA = 6.5)

}
