% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_carbon.R
\name{osi_carbon}
\alias{osi_carbon}
\title{Calculate a carbon index}
\usage{
osi_carbon(B_LU, A_C_OF, B_BGZ, A_CLAY_MI, A_SAND_MI, B_COUNTRY)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{B_BGZ}{(factor) an European region-id used for carbon analyses}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The carbon index. A numeric value.
}
\description{
This function calculates a soil organic carbon index.
}
\examples{
osi_carbon(B_LU = '165',A_C_OF = 25, B_BGZ = '4',A_CLAY_MI=5,A_SAND_MI=25,B_COUNTRY='DE')

}
