% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_helpers.R
\name{osi_checkvar}
\alias{osi_checkvar}
\title{Wrapper function to test euosi input variables for the correct input values or options}
\usage{
osi_checkvar(
  parm,
  fname = NULL,
  na_allowed = FALSE,
  unitcheck = TRUE,
  pwarning = FALSE
)
}
\arguments{
\item{parm}{(list) The osi parameter - value combination to be checked}

\item{fname}{(character) The name of the function where the check is done}

\item{na_allowed}{(character) Are NA allowed in categorial inputs (TRUE or FALSE)}

\item{unitcheck}{(boolean) Option to switch off unit checks (TRUE or FALSE)}

\item{pwarning}{(boolean) Option to print a warning rather than error (stop) message (TRUE or FALSE)}
}
\value{
warning or error messages when parameter value is beyond the allowed range
}
\description{
This function is a wrapper around the function osi_checkvarfun and allows the user to print warning messages
}
