% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_tables.R
\docType{data}
\name{osi_clim}
\alias{osi_clim}
\title{Averaged climatic conditions per country}
\format{
A data.frame with 37 rows and 10 columns:
\describe{
  \item{osi_country}{the country name}
  \item{b_prec_y}{the total annual precipitation, ERA5 (mm)}
  \item{b_prec_sum}{the total summer precipitation, ERA5 (mm)}
  \item{b_prec_win}{the total winter precipitation, ERA5 (mm)}
  \item{b_pet_y}{the total annual evaporation, ERA5 (mm)}
  \item{b_pet_sum}{the total summer evaporation, ERA5 (mm)}
  \item{b_pet_win}{the total winter evaporation, ERA5 (mm)} 
  \item{b_temp_y}{the mean annual temperature, ERA5 (degrees C)}
  \item{b_temp_sum}{the mean summer temperature, ERA5 (degrees C)}
  \item{b_temp_win}{the mean winter temperature, ERA5 (degrees C)} 
}
}
\usage{
osi_clim
}
\description{
Averaged climatic conditions per country
}
\keyword{datasets}
