% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_convmethods.R
\name{osi_conv_magnesium}
\alias{osi_conv_magnesium}
\title{Estimate soil extractable magnesium (-)}
\usage{
osi_conv_magnesium(
  element,
  A_MG_AAA = NA_real_,
  A_MG_AL = NA_real_,
  A_MG_AN = NA_real_,
  A_MG_CC = NA_real_,
  A_MG_CO_PO = NA_real_,
  A_MG_DL = NA_real_,
  A_MG_KCL = NA_real_,
  A_MG_M3 = NA_real_,
  A_MG_NaAAA = NA_real_,
  A_CEC_CO = NA_real_,
  A_PH_CC = NA_real_
)
}
\arguments{
\item{element}{(character) the method requested to be calculated}

\item{A_MG_AAA}{(numeric) The exchangeable K-content of the soil measured via acid ammonium acetate extraction}

\item{A_MG_AL}{(numeric) The exchangeable Mg-content of the soil measured via Ammonium Lactate extraction (mg Mg/ kg)}

\item{A_MG_AN}{(numeric) The Mg-content of the soil extracted with ammonium nitrate (mg Mg /kg)}

\item{A_MG_CC}{(numeric) The plant available content of Mg in the soil (mg  Mg per kg) extracted by 0.01M CaCl2}

\item{A_MG_CO_PO}{(numeric) The exchangeable Mg-content of the soil measured via Cohex extracton, percentage occupation at CEC (\%)}

\item{A_MG_DL}{(numeric) The exchangeable Mg-content of the soil measured via Double Lactate extraction (mg Mg/ kg)}

\item{A_MG_KCL}{(numeric) The plant available potassium, extracted with KCL (mg per kg)}

\item{A_MG_M3}{(numeric) The exchangeable Mg-content of the soil measured via Mehlich 3 extracton (mg Mg/ kg)}

\item{A_MG_NaAAA}{(numeric) The Mg-content of the soil extracted with Morgan's solution, sodium acetate acetic acid (mg/ kg)}

\item{A_CEC_CO}{(numeric) The cation exchange capacity of the soil, measured via Cohenx (mmol+ / kg)}

\item{A_PH_CC}{(numeric) The pH measured in cacl2}
}
\description{
Estimate soil extractable magnesium (-)
}
