% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_convmethods.R
\name{osi_conv_npmn}
\alias{osi_conv_npmn}
\title{Calculate potential mineralizable N for agricultural soils
default values are taken from global ISRIC database}
\usage{
osi_conv_npmn(A_N_RT, A_CLAY_MI, med_PMN = 51.9, med_NRT = 1425, med_CLAY = 26)
}
\arguments{
\item{A_N_RT}{(numeric)  Soil organic N content (mg N / kg) of top soil}

\item{A_CLAY_MI}{(numeric) Clay content (\%)}

\item{med_PMN}{(numeric) Regional median value of PMN (mg N / kg)}

\item{med_NRT}{(numeric) Regional median value of soil organic N content (mg N / kg)}

\item{med_CLAY}{(numeric) Regional median value of clay content (\%)}
}
\description{
Calculate potential mineralizable N for agricultural soils
default values are taken from global ISRIC database
}
