% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_convmethods.R
\name{osi_conv_ph}
\alias{osi_conv_ph}
\title{Estimate soil pH values (-)}
\usage{
osi_conv_ph(
  element,
  A_PH_KCL = NA_real_,
  A_PH_CC = NA_real_,
  A_PH_WA = NA_real_
)
}
\arguments{
\item{element}{(character) the method requested to be calculated}

\item{A_PH_KCL}{(numeric) The soil pH determined in 1 M KCL extract}

\item{A_PH_CC}{(numeric) The soil pH determined in 0.01M CaCl2 extract}

\item{A_PH_WA}{(numeric) The soil pH determined in water}
}
\description{
Estimate soil pH values (-)
}
