% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_convmethods.R
\name{osi_conv_phosphor}
\alias{osi_conv_phosphor}
\title{Estimate soil extractable phosphorus (-)}
\usage{
osi_conv_phosphor(
  element,
  A_P_AL = NA_real_,
  A_P_CC = NA_real_,
  A_P_WA = NA_real_,
  A_P_OL = NA_real_,
  A_P_CAL = NA_real_,
  A_P_DL = NA_real_,
  A_P_AAA = NA_real_,
  A_P_AAA_EDTA = NA_real_,
  A_P_M3 = NA_real_,
  B_SOILTYPE_AGR = NA_real_,
  A_PH_CC = 5
)
}
\arguments{
\item{element}{(character) the method requested to be calculated}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with ammonium lactate}

\item{A_P_CC}{(numeric) The P-content of the soil extracted with CaCl2}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen}

\item{A_P_CAL}{(numeric) The P-content of the soil extracted with ammonium lactate(mg P2O5 / 100g)}

\item{A_P_DL}{(numeric) The P-content of the soil extracted with double lactate (mg P / kg)}

\item{A_P_AAA}{(numeric) The exchangeable P-content of the soil measured via acid ammonium acetate extraction}

\item{A_P_AAA_EDTA}{(numeric) The exchangeable P-content of the soil measured via acid ammonium acetate+EDTA extraction}

\item{A_P_M3}{(numeric) The P-content of the soil extracted with Mehlig 3}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil.}

\item{A_PH_CC}{(numeric) The pH measured in cacl2. If missing, then assume its agronomic common value of 5.}
}
\description{
Estimate soil extractable phosphorus (-)
}
\references{
Steinfurth et al., (2021) Conversion equations between Olsen-P and other methods used to assess plant available soil phosphorus in Europe. A review
}
