% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_convmethods.R
\name{osi_conv_potassium}
\alias{osi_conv_potassium}
\title{Estimate soil extractable potassium (-)}
\usage{
osi_conv_potassium(
  element,
  A_K_AAA = NA_real_,
  A_K_AL = NA_real_,
  A_K_AN = NA_real_,
  A_K_CAL = NA_real_,
  A_K_CC = NA_real_,
  A_K_CO_PO = NA_real_,
  A_K_DL = NA_real_,
  A_K_M3 = NA_real_,
  A_K_NaAAA = NA_real_,
  A_K_WA = NA_real_,
  A_CEC_CO = NA_real_,
  A_PH_CC = NA_real_
)
}
\arguments{
\item{element}{(character) the method requested to be calculated}

\item{A_K_AAA}{(numeric) The exchangeable K-content of the soil measured via acid ammonium acetate extraction}

\item{A_K_AL}{(numeric) The K-content of the soil extracted with ammonium lactate}

\item{A_K_AN}{(numeric) The K-content of the soil extracted with ammonium nitrate (mg K /kg)}

\item{A_K_CAL}{(numeric) The K-content of the soil extracted with ammonium lactate(mg K / kg)}

\item{A_K_CC}{(numeric) The K-content of the soil extracted with CaCl2}

\item{A_K_CO_PO}{(numeric) The occupation of the CEC with potassium (\%)}

\item{A_K_DL}{(numeric) The K-content of the soil extracted with double lactate (mg K / kg)}

\item{A_K_M3}{(numeric) The exchangeable K-content of the soil measured via Mehlich 3 extracton (mg K/ kg)}

\item{A_K_NaAAA}{(numeric) The K-content of the soil extracted with Morgan's solution, sodium acetate acetic acid (mg/ kg)}

\item{A_K_WA}{(numeric) The K-content of the soil extracted with water}

\item{A_CEC_CO}{(numeric) The cation exchange capacity of the soil, measured via Cohenx (mmol+ / kg)}

\item{A_PH_CC}{(numeric) The pH measured in cacl2}
}
\description{
Estimate soil extractable potassium (-)
}
