% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_convmethods.R
\name{osi_conv_som}
\alias{osi_conv_som}
\title{Estimate soil organic matter and nutrient values (-)}
\usage{
osi_conv_som(
  element,
  A_SOM_LOI = NA_real_,
  A_C_OF = NA_real_,
  A_N_RT = NA_real_,
  A_CN_FR = NA_real_
)
}
\arguments{
\item{element}{(character) the method requested to be calculated}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_C_OF}{(numeric) The organic carbon content of the soil in g C / kg}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_CN_FR}{(numeric) The C-to-N ratio of the soil organic matter (-)}
}
\description{
Estimate soil organic matter and nutrient values (-)
}
