% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_convmethods.R
\name{osi_conv_zinc}
\alias{osi_conv_zinc}
\title{Estimate soil extractable zinc (-)}
\usage{
osi_conv_zinc(
  element,
  A_SOM_LOI,
  A_ZN_RT,
  A_PH_CC,
  A_ZN_CC = NA_real_,
  A_ZN_CO = NA_real_,
  A_ZN_DTPA = NA_real_,
  A_ZN_EDTA = NA_real_,
  A_ZN_M3 = NA_real_,
  A_ZN_WA = NA_real_
)
}
\arguments{
\item{element}{(character) the method requested to be calculated}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_ZN_RT}{(numeric) The total Zn-content of the soil via XRF or Dumas (mg Zn/kg)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{A_ZN_CC}{(numeric) The exchangeable Zn-content of the soil measured via 0.01M CaCl2 (ug Zn/kg)}

\item{A_ZN_CO}{(numeric) The exchangeable Zn-content of the soil measured via Cohex extraction (mg Zn/kg)}

\item{A_ZN_DTPA}{(numeric) The exchangeable Zn-content of the soil measured via DTPA (mg Zn/kg)}

\item{A_ZN_EDTA}{(numeric) The exchangeable Zn-content of the soil measured via EDTA (mg Zn/kg)}

\item{A_ZN_M3}{(numeric) The exchangeable Zn-content of the soil measured via Mehlich-III (mg Zn/kg)}

\item{A_ZN_WA}{(numeric) The exchangeable Zn-content of the soil measured via water (mg Zn/kg)}
}
\description{
Estimate soil extractable zinc (-)
}
