% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_tables.R
\docType{data}
\name{osi_countries}
\alias{osi_countries}
\title{Table with country codes used in the package}
\format{
A data.table with 249 rows and 3 columns:
\describe{
  \item{country_name}{the name of the country}
  \item{country_code}{the country code;from https://datahub.io/core/country-codes}
}
}
\usage{
osi_countries
}
\description{
This table contains country names as used for the spatial extent where sptfs have been derived from.
}
\keyword{datasets}
