% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_wat_erosion.R
\name{osi_erosion}
\alias{osi_erosion}
\title{Calculate the water erosion index using RUSLE approach}
\usage{
osi_erosion(B_LU, A_SOM_LOI, A_CLAY_MI, A_SAND_MI, B_COUNTRY)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The water erosion index. A numeric value.
}
\description{
This function calculates the water erosion index
}
\examples{
osi_erosion(B_LU = 'SOJ',A_SOM_LOI = 3.5,A_CLAY_MI = 5,A_SAND_MI = 15,B_COUNTRY='NL')

}
