% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_evaluate.R
\name{osi_evaluate_parabolic}
\alias{osi_evaluate_parabolic}
\title{Evaluate a soil quality function using a parabolic function}
\usage{
osi_evaluate_parabolic(x, x.top, must.plateau = TRUE)
}
\arguments{
\item{x}{(numeric) The values of a calc function to be converted to an evaluation}

\item{x.top}{(numeric) The value at which x reaches the plateau}

\item{must.plateau}{(boolean) set max at plateau or accept decline after reaching optimum.}
}
\value{
A transformed variable after applying a parabolic evaluation function. A numeric value.
}
\description{
This function evaluates the calculated values from an indicator using a parabolic function. After the optimum is reached the it stays at its plateau.
}
\examples{
osi_evaluate_parabolic(x = 5, x.top = 8)
osi_evaluate_parabolic(x = c(0.1,0.5,1.5,3.5), x.top = 6.5)

}
