% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_tables.R
\docType{data}
\name{osi_farm}
\alias{osi_farm}
\title{Example farm database to illustrate the functionality of the euosi package}
\format{
A data.frame with 60 rows and 24 columns:
\describe{
  \item{ID}{the unique field-id}
  \item{B_COUNTRY}{the country name}
  \item{B_LU}{The BRP gewascode of the crop}
  \item{B_PREC_SUM}{the total summer precipitation, ERA5 (mm)}
  \item{B_PREC_WIN}{the total winter precipitation, ERA5 (mm)}
  \item{B_PET_SUM}{the total summer evaporation, ERA5 (mm)}
  \item{B_PET_WIN}{the total winter evaporation, ERA5 (mm)} 
  \item{B_TEMP_SUM}{the mean summer temperature, ERA5 (degrees C)}
  \item{B_TEMP_WIN}{the mean winter temperature, ERA5 (degrees C)} 
  \item{A_SOM_LOI}{The percentage organic matter in the soil (\%)}
  \item{A_CLAY_MI}{The clay content of the soil (\%)}
  \item{A_SAND_MI}{The sand content of the soil (\%)}
  \item{A_PH_CC}{The acidity of the soil, measured in 0.01M CaCl2 (-)}
  \item{A_CACO3_IF}{the percentage of CaCO3 (\%)}
  \item{A_CEC_CO}{The cation exchange capacity of the soil (mmol+ / kg), analyzed via Cobalt-hexamine extraction}
  \item{A_C_OF}{The carbon content of the soil layer (g/ kg)}
  \item{A_N_RT}{The organic nitrogen content of the soil in mg N / kg}
  \item{A_N_PMN}{The potentially mineralizable N pool (mg N / kg soil)}
  \item{A_P_OL}{The P-content of the soil extracted with Olsen (mg P / kg)}
  \item{A_K_AAA}{The exchangeable K-content of the soil measured via ammonium acetate extraction (mg K / kg)}
  \item{A_B_HW}{The plant available content of B in the soil extracted by hot water (mg  B / kg)}
  \item{A_ZN_CC}{The plant available content of Zn in the soil  extracted by CaCl2 (mg Zn / kg)} 
  \item{A_MG_AAA}{The exchangeable Mg-content of the soil measured via ammonium acetate extraction (mg Mg / kg)}          
}
}
\usage{
osi_farm
}
\description{
Example farm database to illustrate the functionality of the euosi package
}
\keyword{datasets}
