% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_main.R
\name{osi_field_dt}
\alias{osi_field_dt}
\title{Calculate the OSI score for one field using a data.table as input}
\usage{
osi_field_dt(dt, output = "all")
}
\arguments{
\item{dt}{(data.table) A data.table with all input required to calculate BLN on field level}

\item{output}{(character) An optional argument to select output: scores, indicators, or all. (default = all)}
}
\description{
This functions wraps the functions of the euosi into one main function to calculate the soil quality score for a single field. Whereas the function `osi_field` requires all variables as separate inputs, the function `osi_field_dt` allows one to send in a data.table
}
