% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_helpers.R
\name{osi_get_SOILTYPE_AGR}
\alias{osi_get_SOILTYPE_AGR}
\title{Estimate agricultural soil type classification used in the Netherlands}
\usage{
osi_get_SOILTYPE_AGR(A_CLAY_MI, A_SAND_MI, A_SOM_LOI, A_PH_CC)
}
\arguments{
\item{A_CLAY_MI}{(numeric) Clay content (\%)}

\item{A_SAND_MI}{(numeric) Silt content (\%)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}
}
\value{
agricultural soil type according to the Dutch classification system
}
\description{
Estimate agricultural soil type classification used in the Netherlands
}
