% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_helpers.R
\name{osi_get_TEXTURE_BE}
\alias{osi_get_TEXTURE_BE}
\title{Estimate soil texture according to Belgium classification}
\usage{
osi_get_TEXTURE_BE(A_CLAY_MI, A_SILT_MI, A_SAND_MI, type = "code")
}
\arguments{
\item{A_CLAY_MI}{(numeric) Clay content (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SAND_MI}{(numeric) Silt content (\%)}

\item{type}{(character) return Belgium classification names or codes (options: 'name' or 'code')}
}
\value{
Texture class according to the Belgium classification system
}
\description{
Estimate soil texture according to Belgium classification
}
