% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen_loss.R
\name{osi_gw_nleach}
\alias{osi_gw_nleach}
\title{Calculate the risk for nitrogen leaching in Europe}
\usage{
osi_gw_nleach(
  B_LU = NA_character_,
  A_CLAY_MI = NA_real_,
  A_SAND_MI = NA_real_,
  A_CACO3_IF = NA_real_,
  A_N_RT = NA_real_,
  A_C_OF = NA_real_,
  B_PREC_SUM = NA_real_,
  B_PREC_WIN = NA_real_,
  B_PET_SUM = NA_real_,
  B_PET_WIN = NA_real_,
  B_TEMP_SUM = NA_real_,
  B_TEMP_WIN = NA_real_,
  B_COUNTRY,
  pwarning = FALSE
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_CACO3_IF}{(numeric) The percentage of carbonated lime (\%)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{B_PREC_SUM}{(numeric) Total potential precipitation in summer (mm)}

\item{B_PREC_WIN}{(numeric) Total potential precipitation in winter (mm)}

\item{B_PET_SUM}{(numeric) Total potential evapotranspiration in summer (mm)}

\item{B_PET_WIN}{(numeric) Total potential evapotranspiration in winter (mm)}

\item{B_TEMP_SUM}{(numeric) Mean winter temperature (degrees Celcius)}

\item{B_TEMP_WIN}{(numeric) Mean winter temperature (degrees Celcius)}

\item{B_COUNTRY}{(character) The country code}

\item{pwarning}{(boolean) Option to print a warning rather than error (stop) message for input checks (TRUE or FALSE)}
}
\value{
The risk for N leaching from agricultural soils in Europe
}
\description{
This function calculates the N leaching risks for the soils in Europe
}
