% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen_loss.R
\name{osi_gw_nleach_eu}
\alias{osi_gw_nleach_eu}
\title{Calculate the risk for nitrogen leaching template for EU}
\usage{
osi_gw_nleach_eu(
  B_LU,
  A_N_RT,
  A_C_OF,
  A_CLAY_MI,
  A_SAND_MI,
  B_PREC_Y = NA_real_,
  B_PET_Y = NA_real_,
  B_TEMP_Y = NA_real_,
  B_COUNTRY
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{B_PREC_Y}{(numeric) Total potential precipitation in year (mm)}

\item{B_PET_Y}{(numeric) Total potential evapotranspiration in year (mm)}

\item{B_TEMP_Y}{(numeric) Mean year temperature (degrees Celcius)}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The risk for N leaching
}
\description{
This function calculates the N leaching risks for the soil
}
\examples{
osi_gw_nleach_eu(B_LU ='265',A_N_RT = 4100, A_C_OF = 22, A_CLAY_MI = 4.5, 
A_SAND_MI = 15, B_PREC_Y = 900, B_PET_Y = 650,B_TEMP_Y = 4.5)

}
