% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen_loss.R
\name{osi_gw_nleach_fi}
\alias{osi_gw_nleach_fi}
\title{Calculate the risk for nitrogen leaching in Finland}
\usage{
osi_gw_nleach_fi(B_LU, A_N_RT, A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}
}
\value{
The risk for N leaching
}
\description{
This function calculates the N leaching risks for the soil
}
\examples{
osi_gw_nleach_fi(B_LU = '772',A_N_RT = 1200, A_C_OF = 25, A_CLAY_MI = 3.5, A_SAND_MI = 15)

}
