% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nitrogen_loss.R
\name{osi_gw_nleach_fr}
\alias{osi_gw_nleach_fr}
\title{Calculate the risk for nitrogen leaching in France}
\usage{
osi_gw_nleach_fr(
  B_LU,
  A_N_RT,
  A_C_OF,
  A_CLAY_MI,
  A_SAND_MI,
  A_CACO3_IF,
  B_PREC_SUM = NA_real_,
  B_PREC_WIN = NA_real_,
  B_PET_SUM = NA_real_,
  B_PET_WIN = NA_real_,
  B_TEMP_SUM = NA_real_,
  B_TEMP_WIN = NA_real_
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_CACO3_IF}{(numeric) The percentage of carbonated lime (\%)}

\item{B_PREC_SUM}{(numeric) Total potential precipitation in summer (mm)}

\item{B_PREC_WIN}{(numeric) Total potential precipitation in winter (mm)}

\item{B_PET_SUM}{(numeric) Total potential evapotranspiration in summer (mm)}

\item{B_PET_WIN}{(numeric) Total potential evapotranspiration in winter (mm)}

\item{B_TEMP_SUM}{(numeric) Mean winter temperature (degrees Celcius)}

\item{B_TEMP_WIN}{(numeric) Mean winter temperature (degrees Celcius)}
}
\value{
The risk of nitrogen leaching. A numeric value, converted to a OSI score.
}
\description{
This function calculates the N leaching risks for the soil
}
\examples{
osi_gw_nleach_fr(B_LU = 'BDH',A_N_RT = 1200, A_C_OF = 25, 
A_CLAY_MI = 3.5, A_SAND_MI = 15,A_CACO3_IF = 0.8)

}
