% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_at}
\alias{osi_nut_k_at}
\title{Calculate the potassium excess index in Austria}
\usage{
osi_nut_k_at(A_K_CAL, B_TEXTURE_HYPRES, B_LU = NA_character_)
}
\arguments{
\item{A_K_CAL}{(numeric) The exchangeable K-content of the soil measured via Calcium Ammonium Lactate (mg K/ kg)}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium excess index in Austria estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_at(B_LU = '3301000000',A_K_CAL = 47,B_TEXTURE_HYPRES = 'C')

}
