% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_ch}
\alias{osi_nut_k_ch}
\title{Calculate the potassium excess index in Switzerland}
\usage{
osi_nut_k_ch(A_K_AAA, A_CLAY_MI, B_LU = NA_character_)
}
\arguments{
\item{A_K_AAA}{(numeric) The exchangeable K-content of the soil measured via acid ammonium acetate extraction (mg K/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium excess index in Switzerland estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_ch(B_LU = '8410',A_K_AAA = 50, A_CLAY_MI = 4.5)

}
