% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_cz}
\alias{osi_nut_k_cz}
\title{Calculate the potassium excess index in Czech Republic}
\usage{
osi_nut_k_cz(A_K_M3, B_TEXTURE_HYPRES, B_LU = NA_character_)
}
\arguments{
\item{A_K_M3}{(numeric) The exchangeable K-content of the soil measured via Mehlich 3 extracton (mg K/ kg)}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium excess index in Czech Republic estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_cz(B_LU = '3301000000', A_K_M3 = 81, B_TEXTURE_HYPRES='C')

}
