% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_de}
\alias{osi_nut_k_de}
\title{Calculate the potassium excess index in Germany}
\usage{
osi_nut_k_de(B_LU, A_C_OF, A_CLAY_MI, A_SAND_MI, A_K_CAL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_C_OF}{(numeric) The carbon content of the soil layer (g/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_K_CAL}{(numeric) The potassium content extracted with calcium ammonium lactate (g / kg)}
}
\value{
The potassium excess index in Germany derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess.
}
