% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_dk}
\alias{osi_nut_k_dk}
\title{Calculate the potassium excess index in Denmark}
\usage{
osi_nut_k_dk(B_LU, A_K_AL, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_K_AL}{(numeric) The K-content of the soil extracted with ammonium lactate (mg K / kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium excess index in Denmark derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_dk(B_LU = '3301000000',A_K_AL = 5)
osi_nut_k_dk(B_LU = c('3301000000','3301061299'),A_K_AL = c(3.5,5.5))

}
