% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_es}
\alias{osi_nut_k_es}
\title{Calculate the potassium excess index in Spain}
\usage{
osi_nut_k_es(B_LU, B_TEXTURE_HYPRES, A_K_AAA, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{A_K_AAA}{(numeric) The K-content of the soil extracted with ammoninium acetate (mg K/kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium excess index in Spain derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_es(B_LU = '3301000000',A_K_AAA = 5,B_TEXTURE_HYPRES='C')

}
