% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_fr}
\alias{osi_nut_k_fr}
\title{Calculate the potassium excess index in France}
\usage{
osi_nut_k_fr(
  B_LU,
  A_K_AAA,
  B_TEXTURE_GEPPA = NA_character_,
  B_SOILTYPE_AGR = NA_character_,
  B_AER_FR = NA_character_,
  A_PH_WA = NA_real_
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_K_AAA}{(numeric) The exchangeable K-content of the soil measured via ammonium acetate extraction}

\item{B_TEXTURE_GEPPA}{(character) The soil texture class in a particular region.}

\item{B_SOILTYPE_AGR}{(character) The agricultural soil type classification}

\item{B_AER_FR}{(character) An agroeconomic region in France. Optional argument.}

\item{A_PH_WA}{(numeric) The pH measured in water.}
}
\value{
The potassium excess index in France estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\details{
The function has two optional arguments soil type (B_SOILTYPE_AGR) and agricultural region (B_AER_FR). When these are unknown, then the soil type is estimated based on the pH value. Threshold values are then generalized for calcareous and non-calcareous soils.
}
\examples{
osi_nut_k_fr(B_LU = 'SOJ', A_K_AAA = 45,B_TEXTURE_GEPPA ='As',
B_SOILTYPE_AGR = 'limons battants', B_AER_FR = 'nord picardie')

}
