% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_ie}
\alias{osi_nut_k_ie}
\title{Calculate the potassium excess index in Ireland}
\usage{
osi_nut_k_ie(B_LU, A_SOM_LOI, A_K_NaAAA, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_K_NaAAA}{(numeric) The K-content of the soil extracted with Morgan's solution, sodium acetate acetic acid (mg/ kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium excess index in Ireland derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_ie(B_LU = 'testcrop',A_SOM_LOI = 2,A_K_NaAAA = 5)
osi_nut_k_ie(B_LU = c('testcrop','testcrop2'),A_SOM_LOI = c(2,4),A_K_NaAAA = c(3.5,5.5))

}
