% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_no}
\alias{osi_nut_k_no}
\title{Calculate the potassium excess index in Norway}
\usage{
osi_nut_k_no(B_LU, A_K_AL, A_CLAY_MI)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_K_AL}{(numeric) The K-content of the soil extracted with ammonium lactate (mg K / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}
}
\value{
The potassium excess index in Norway derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_no(B_LU = 265,A_K_AL = 5,A_CLAY_MI=5)
osi_nut_k_no(B_LU = c(265,1019),A_K_AL = c(3.5,5.5),A_CLAY_MI=c(3,5))

}
