% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_pt}
\alias{osi_nut_k_pt}
\title{Calculate the potassium excess index index in Portugal}
\usage{
osi_nut_k_pt(B_LU, A_K_AAA, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_K_AAA}{(numeric) The K-content of the soil extracted with acid ammonium acetate (mg K / kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium availability index in Portugal derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess index.
}
\examples{
osi_nut_k_pt(B_LU = '3301061299',A_K_AAA = 50)
osi_nut_k_pt(B_LU =  c('3301061299','3301000000'),A_K_AAA = c(35,55))

}
