% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_ro}
\alias{osi_nut_k_ro}
\title{Calculate thepotassium excess index in Romenia}
\usage{
osi_nut_k_ro(A_K_AL, B_TEXTURE_HYPRES, B_LU = NA_character_)
}
\arguments{
\item{A_K_AL}{(numeric) The exchangeable K-content of the soil measured via ammonium lactate extracton (mg K/ kg)}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium excess index in Romenia estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium excess index
}
\examples{
osi_nut_k_ro(B_LU = 'testcrop1',A_K_AL = 45,B_TEXTURE_HYPRES='M')

}
