% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_se}
\alias{osi_nut_k_se}
\title{Calculate the potassium excess index in Sweden}
\usage{
osi_nut_k_se(B_LU, A_K_AL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_K_AL}{(numeric) The K-content of the soil extracted with ammonium lactate (mg K / kg)}
}
\value{
The potassium excess index in Sweden derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_se(B_LU = '3301061299',A_K_AL = 5)
osi_nut_k_se(B_LU =  c('3301061299','3301000000'),A_K_AL = c(3.5,5.5))

}
