% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_sk}
\alias{osi_nut_k_sk}
\title{Calculate the potassium excess index in  Slovak Republic}
\usage{
osi_nut_k_sk(B_LU, B_TEXTURE_HYPRES, A_K_M3, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{A_K_M3}{(numeric) The exchangeable K-content of the soil measured via Mehlich 3 extracton (mg K/ kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium excess index in Slovak Republic estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_sk(B_LU = '3301010901', A_K_M3 = 45,B_TEXTURE_HYPRES='C')

}
