% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_uk}
\alias{osi_nut_k_uk}
\title{Calculate the potassium excess index in United Kingdom}
\usage{
osi_nut_k_uk(B_LU, A_SOM_LOI, A_K_AN, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_K_AN}{(numeric) The K-content of the soil extracted with ammonium nitrate (mg K /kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium excess index in United Kingdom derived from extractable soil K fractions. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_uk(B_LU = 'testcrop1',A_SOM_LOI=3,A_K_AN = 50)
osi_nut_k_uk(B_LU = c('testcrop1','testcrop2'),A_SOM_LOI = c(3,5),A_K_AN = c(35,55))

}
