% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_at}
\alias{osi_nut_p_at}
\title{Calculate the phosphorus excess index in Austria}
\usage{
osi_nut_p_at(A_P_CAL, B_LU = NA_character_)
}
\arguments{
\item{A_P_CAL}{(numeric) The exchangeable P-content of the soil measured via Calcium Ammonium Lactate (mg P/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus excess index in Austria estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess
}
\examples{
osi_nut_p_at(A_P_CAL = 47)

}
