% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_be}
\alias{osi_nut_p_be}
\title{Calculate the phosphorus excess index in Belgium}
\usage{
osi_nut_p_be(B_LU, A_P_AL)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_P_AL}{(numeric) The exchangeable P-content of the soil measured via ammonium lactate extraction}
}
\value{
The phosphorus excess index in Belgium estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess index
}
\examples{
osi_nut_p_be(B_LU = 'SOJ', A_P_AL = 45)

}
