% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_ch}
\alias{osi_nut_p_ch}
\title{Calculate the phosphorus excess index in Switzerland}
\usage{
osi_nut_p_ch(A_P_AAA, B_LU = NA_character_)
}
\arguments{
\item{A_P_AAA}{(numeric) The exchangeable P-content of the soil measured via acid ammonium acetate extraction (mg P/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus excess index in Switzerland estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess
}
\examples{
osi_nut_p_ch(A_P_AAA = 50)

}
