% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_de}
\alias{osi_nut_p_de}
\title{Calculate the phosphate excess index in Germany}
\usage{
osi_nut_p_de(
  B_LU,
  A_SOM_LOI,
  A_P_CAL = NA_real_,
  A_P_DL = NA_real_,
  unitcheck = TRUE
)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_P_CAL}{(numeric) The P-content of the soil extracted with ammonium lactate(mg P / kg)}

\item{A_P_DL}{(numeric) The P-content of the soil extracted with double lactate (mg P / kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphate excess index in Germany stimated from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess
}
