% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_dk}
\alias{osi_nut_p_dk}
\title{Calculate the phosphate excess index in Denmark}
\usage{
osi_nut_p_dk(B_LU, A_P_OL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg/kg)}
}
\value{
The phosphate excess index in Denmark derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess
}
\examples{
osi_nut_p_dk(B_LU = '3301010399',A_P_OL = 5)
osi_nut_p_dk(B_LU = c('3301010399','3301029800'),A_P_OL = c(3.5,5.5))

}
