% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_ee}
\alias{osi_nut_p_ee}
\title{Calculate the phosphorus excess index in Estonia}
\usage{
osi_nut_p_ee(A_P_M3, A_SOM_LOI, B_LU = NA_character_)
}
\arguments{
\item{A_P_M3}{(numeric) The exchangeable P-content of the soil measured via Mehlich 3 extracton (mg P/ kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus excess index in Estonia estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess
}
\examples{
osi_nut_p_ee(A_P_M3 = 45,A_SOM_LOI = 3, B_LU='testcrop1')

}
