% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_el}
\alias{osi_nut_p_el}
\title{Calculate the phosphate excess index in Greece}
\usage{
osi_nut_p_el(B_LU, A_P_OL, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg P/kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphate availability index in Greece derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess index.
}
\examples{
osi_nut_p_el(B_LU = '3301061299',A_P_OL = 5)
osi_nut_p_el(B_LU = c('3301061299','3301000000'),A_P_OL = c(5,15))

}
