% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_fi}
\alias{osi_nut_p_fi}
\title{Calculate the phosphorus excess index in Finland}
\usage{
osi_nut_p_fi(B_LU, B_TEXTURE_USDA, A_P_AAA, A_C_OF = 0)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_TEXTURE_USDA}{(character) The soil texture according to USDA classification system}

\item{A_C_OF}{(numeric) The organic carbon content in the soil (g C / kg)}

\item{A_P_AAA(numeric)}{The exchangeable P-content of the soil measured via ammonium acetate extraction}
}
\value{
The phosphorus excess index in Finland estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess index
}
\examples{
osi_nut_posphor_fi(B_LU = 'SOJ', B_TEXTURE_USDA = 'Si',A_P_AAA= 45)

}
