% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_fr}
\alias{osi_nut_p_fr}
\title{Calculate the phosphorus excess index in France}
\usage{
osi_nut_p_fr(B_LU, A_P_OL, A_PH_WA = NA_real_, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen}

\item{A_PH_WA}{(numeric) The pH measured in water.}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphate excess index in France estimated from extractable soil P Olsen (a numeric value).
}
\description{
This function calculates the phosphorus excess index
}
\details{
This function does not account for variability per agricultural region. The pH is used to classify wether soils are calcareous.
}
\examples{
osi_nut_p_fr(B_LU = 'SOJ', A_P_OL = 45, A_PH_WA = 4.5)

}
