% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_ie}
\alias{osi_nut_p_ie}
\title{Calculate the phosphate excess index in Ireland}
\usage{
osi_nut_p_ie(B_LU, A_P_OL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg/kg)}
}
\value{
The phosphate excess index in Ireland derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess.
}
\examples{
osi_nut_p_ie(B_LU = 'testcrop1',A_P_OL = 5)
osi_nut_p_ie(B_LU = c('testcrop1','testcrop2'),A_P_OL = c(3.5,5.5))

}
