% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_lt}
\alias{osi_nut_p_lt}
\title{Calculate the phosphorus excess index in Lithuania}
\usage{
osi_nut_p_lt(A_P_AL, A_SOM_LOI, B_LU = NA_character_)
}
\arguments{
\item{A_P_AL}{(numeric) The exchangeable P-content of the soil measured via Ammonium Lactate extraction (mg P/ kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus excess index in Lithuania estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess.
}
\examples{
osi_nut_p_lt(B_LU = 'testcrop1', A_P_AL = 45,A_SOM_LOI = 4.5)

}
