% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_no}
\alias{osi_nut_p_no}
\title{Calculate the phosphorus excess index in Norway}
\usage{
osi_nut_p_no(A_P_AL, B_LU = NA_character_)
}
\arguments{
\item{A_P_AL}{(numeric) The exchangeable P-content of the soil measured via Ammonium Lactate (mg P/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus excess index in Norway estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess.
}
\examples{
osi_nut_p_no(B_LU = 'testcrop1', A_P_AL = 50)

}
