% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_pl}
\alias{osi_nut_p_pl}
\title{Calculate the phosphorus excess index in Poland}
\usage{
osi_nut_p_pl(A_P_DL, B_LU = NA_character_)
}
\arguments{
\item{A_P_DL}{(numeric) The exchangeable P-content of the soil measured via ammonium double lactate extracton (mg P/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus excess index in Poland estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess.
}
\examples{
osi_c_phosphorus_pl(A_P_DL = 45)

}
