% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_pt}
\alias{osi_nut_p_pt}
\title{Calculate the phosphate excess index in Portugal}
\usage{
osi_nut_p_pt(B_LU, A_P_OL, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg P/kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphate availability index in Portugal derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess.
}
\examples{
osi_nut_p_pt(B_LU = '3301061299',A_P_OL = 5)
osi_nut_p_pt(B_LU = c('3301061299','3301000000'),A_P_OL = c(3.5,5.5))

}
