% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_ro}
\alias{osi_nut_p_ro}
\title{Calculate the phosphate excess index index in Romenia}
\usage{
osi_nut_p_ro(B_LU, A_P_AL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with Acetate Lactate (mg/kg)}
}
\value{
The phosphate availability index in Romenia derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess index.
}
\examples{
osi_c_phosphor_ro(B_LU = 'testcrop1',A_P_AL = 5)
osi_c_phosphor_ro(B_LU = c('testcrop1','testcrop2'),A_P_AL = c(3.5,5.5))

}
