% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_se}
\alias{osi_nut_p_se}
\title{Calculate the phosphate excess index in Sweden}
\usage{
osi_nut_p_se(B_LU, A_P_AL)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with ammonium lactate (mg P / kg)}
}
\value{
The phosphate excess index in Sweden derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess.
}
\examples{
osi_nut_p_se(B_LU = '3301061299',A_P_AL = 5)
osi_nut_p_se(B_LU = c('3301061299','3301000000'),A_P_AL = c(3.5,5.5))

}
