% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_sl}
\alias{osi_nut_p_sl}
\title{Calculate the phosphorus excess index in Slovenia}
\usage{
osi_nut_p_sl(A_P_AL, B_LU = NA_character_)
}
\arguments{
\item{A_P_AL}{(numeric) The exchangeable P-content of the soil measured via ammoniuml lactate extracton (mg P/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus excess index in Slovenia estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus excess.
}
\examples{
osi_c_phosphorus_sl(A_P_AL = 45)

}
