% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_phosphorus.R
\name{osi_nut_p_uk}
\alias{osi_nut_p_uk}
\title{Calculate the phosphate excess index in United Kingdom}
\usage{
osi_nut_p_uk(B_LU, A_SOM_LOI, A_P_OL, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg P /kg)}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The phosphate excess index in United Kingdom derived from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate excess.
}
\examples{
osi_nut_p_uk(B_LU = 'testcrop1',A_SOM_LOI = 3,A_P_OL = 5)
osi_nut_p_uk(B_LU = c('testcrop1','testcrop2'),A_SOM_LOI = c(3,4),A_P_OL = c(3.5,5.5))

}
