% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_crumbleability.R
\name{osi_p_crumbleability}
\alias{osi_p_crumbleability}
\title{Calculate crumbleability index in the topsoil for all EU countries}
\usage{
osi_p_crumbleability(B_LU, A_CLAY_MI, A_SOM_LOI, A_PH_CC, B_COUNTRY)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_PH_CC}{(numeric) The pH measured in CaCl2}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The function returns the crumbleability index
}
\description{
This function calculates the crumbleability index
}
\examples{
osi_p_crumbleability(B_LU = '256', A_CLAY_MI = 5, A_SOM_LOI = 3.5, A_PH_CC = 5.5, B_COUNTRY = 'NL')

}
