% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_density.R
\name{osi_p_density}
\alias{osi_p_density}
\title{Calculate the bulk density}
\usage{
osi_p_density(A_SOM_LOI, A_CLAY_MI, pwarning = FALSE)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{pwarning}{(boolean) Option to print a warning rather than error (stop) message for input checks (TRUE or FALSE)}
}
\value{
The bulk density of an arable soil (kg / m3) evaluated given the maximum density that limit the root penetration.
}
\description{
This pedotransfer function calculates the bulk density of the soil based on texture and organic matter
}
\examples{
osi_p_density(A_SOM_LOI = 6.5, A_CLAY_MI = 28)
osi_p_density(A_SOM_LOI = 3.5, A_CLAY_MI = 2)
osi_p_density(A_SOM_LOI = c(3.5,8.5),A_CLAY_MI = c(2,28))

}
