% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_plant_available_water.R
\name{osi_p_paw}
\alias{osi_p_paw}
\title{Calculate indicator for plant available water}
\usage{
osi_p_paw(A_CLAY_MI, A_SAND_MI, A_SILT_MI, A_C_OF)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_C_OF}{(numeric) The organic carbon content of the soil (g / kg)}
}
\description{
This function calculates the plant available water index.
}
\examples{
osi_p_paw(A_CLAY_MI = 4.5, A_SAND_MI = 23, A_SILT_MI = 72.5,A_C_OF = 23)

}
