% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_plot.R
\name{osi_plot_shi}
\alias{osi_plot_shi}
\title{This function calculates the phosphate availability.}
\usage{
osi_plot_shi(shi, xpar, xpar_label = "soil indicator", pgroup = NULL)
}
\arguments{
\item{shi}{(numeric) A soil health indicator value estimated with euosi}

\item{xpar}{(numeric) The soil property controlling the shi value}

\item{xpar_label}{(character) The label to be printed on the x-axis of the plot}

\item{pgroup}{(character) A grouping factor to show different responses of shi}
}
\description{
This function calculates the phosphate availability.
}
