% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_tables.R
\docType{data}
\name{osi_thresholds}
\alias{osi_thresholds}
\title{Table with country specific threshold values to evaluate soilquality'
 
This table contains the evaluation coefficients to convert a soil function value into a soil quality score ranging from 0 to 1. 
The scoring function migh vary per country, indicator and subcatogries defined by variable categories of soil and crop types.}
\format{
A data.table with x rows and 12 columns:
\describe{
  \item{osi_trh_id}{the threshold category id}
  \item{osi_country}{the country name where the threshold coefficients apply}
  \item{osi_esd}{the soil ecosystem service for which the threshold applies}
  \item{osi_category}{the soil quality category: chemical, biological, physical, other}
  \item{osi_indicator}{the osi indicator for which the threshold applies}
  \item{osi_indicator_name}{the osi indicator name}
  \item{osi_threshold_cropcat}{the osi crop category for which specific thresholds apply. if empy, then the threshold is generally applicable}
  \item{osi_threshold_soilcat}{the osi soil category for which specific thresholds apply. if empy, then the threshold is generally applicable}
  \item{osi_scoringtype}{the soil evaluation function to score the soil function. Options: parabolic, logistic, linear}
  \item{osi_st_c1}{the first coefficient of the scoring function selected}
  \item{osi_st_c2}{the second coefficient of the scoring function selected}
  \item{osi_st_c3}{the third coefficient of the scoring function selected}
}
}
\usage{
osi_thresholds
}
\description{
Table with country specific threshold values to evaluate soilquality'
 
This table contains the evaluation coefficients to convert a soil function value into a soil quality score ranging from 0 to 1. 
The scoring function migh vary per country, indicator and subcatogries defined by variable categories of soil and crop types.
}
\keyword{datasets}
