% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_tables.R
\docType{data}
\name{osi_vars_input}
\alias{osi_vars_input}
\title{A dataset with all input variables required by the euosi package}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 150 rows and 7 columns.
}
\usage{
osi_vars_input
}
\description{
This table contains all input parameters needed to run the euosi package
}
\details{
\describe{
  \item{code}{The input parameter's name}
  \item{parameter}{A description of the parameter}
  \item{unit}{The unit for the parameter}
  \item{data_type}{Type of data, either boolean, numeric, integer, or character}
  \item{value_min}{Minimum value for a numeric or integer parameter}
  \item{value_max}{Maximum value for a numeric or integer parameter}
  \item{categories}{Possible categories for categorial variables}
}
}
\keyword{datasets}
