% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_tables.R
\docType{data}
\name{osi_vars_output}
\alias{osi_vars_output}
\title{A dataset with all output variables generated by the euosi package}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 29 rows and 2 columns.
}
\usage{
osi_vars_output
}
\description{
This table contains all soil indicators and soil quality scores for the euosi package
}
\details{
\describe{
  \item{code}{The name of the output parameter}
  \item{parameter}{A description of the output parameter}
}
}
\keyword{datasets}
