% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_evaluate.R
\name{cf_ind_importance}
\alias{cf_ind_importance}
\title{Helper function to weight and correct the risk and scores}
\usage{
cf_ind_importance(value)
}
\arguments{
\item{value}{The risk or OSI core value to be weighted}
}
\value{
A transformed variable after applying a inverse weighing function so that lower values will gain more impact when applied in a weighed.mean function. A numeric value.
}
\description{
Helper function to weight and correct the risk and scores
}
\examples{
cf_ind_importance(value = 0.5)
cf_ind_importance(value = c(0.1,0.5,1.5))

}
