% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_boron.R
\name{osi_c_boron}
\alias{osi_c_boron}
\title{Calculate the boron availability index (wrapper function)}
\usage{
osi_c_boron(A_CLAY_MI, A_BO_HW, B_COUNTRY, B_LU)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content %}

\item{A_BO_HW}{(numeric) The plant available content of B in the soil (mg  Cu per kg) extracted by hot water}

\item{B_LU}{(numeric) The crop code}
}
\value{
The capacity of the soil to supply and buffer boron, evaluated given an optimum threshold for yield. A numeric value.
}
\description{
This function calculates the boron availability for all European countries (if available).
}
