% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_copper.R
\name{osi_c_copper_fr}
\alias{osi_c_copper_fr}
\title{Calculate the Cu availability index for the Netherlands}
\usage{
osi_c_copper_fr(A_CLAY_MI, A_CU_ED, A_SOM_LOI, B_CF)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content %}

\item{A_CU_ED}{(numeric) The plant available content of Cu in the soil (mg Cu per kg) extracted by EDTA}

\item{A_SOM_LOI}{(numeric) The organic matter content %}

\item{B_CF}{(numeric) The % of coarse fragments}
}
\value{
The copper availability index in France estimated from extractable copper, clay, coarse fragments and soil organic matter, A numeric value.
}
\description{
This function calculates the Cu availability of a soil, using the agronomic index used in the Netherlands.
}
\details{
#' Calculate the Cu availability index for France 

This function calculates the Cu availability of a soil, using the agronomic index used in France
}
