% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_es}
\alias{osi_c_magnesium_es}
\title{Calculate the magnesium availability index in Spain}
\usage{
osi_c_magnesium_es(B_LU, A_MG_CO_PO)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_MG_CO_PO}{(numeric) The exchangeable Mg-content of the soil measured via Cohex extracton, percentage occupation at CEC (\%)}
}
\value{
The magnesium availability index in Spain estimated from the magnesium occupation at CEC. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_es(B_LU = '3301000000' , A_MG_CO_PO = 4.5)

}
