% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_magnesium.R
\name{osi_c_magnesium_it}
\alias{osi_c_magnesium_it}
\title{Calculate the magnesium availability index in Italy}
\usage{
osi_c_magnesium_it(A_MG_CO_PO, A_K_CO_PO, B_LU = NA_character_)
}
\arguments{
\item{A_MG_CO_PO}{(numeric) The exchangeable Mg-content of the soil measured via Cohex extracton, percentage occupation at CEC (\%)}

\item{A_K_CO_PO}{(numeric) The exchangeable K-content of the soil measured via Cohex extracton, percentage occupation at CEC (\%)}

\item{B_LU}{(character) The crop code}
}
\value{
The magnesium availability index in Italy estimated from the Mg extracted soil pool, expressed as occupation of the CEC. A numeric value.
}
\description{
This function calculates the magnesium availability.
}
\examples{
osi_c_magnesium_it(A_MG_CO_PO = 4.5,A_K_CO_PO = 3)

}
