% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_ph.R
\name{osi_c_ph_fr}
\alias{osi_c_ph_fr}
\title{Calculate the pH index for France}
\usage{
osi_c_ph_fr(B_LU, B_TEXTURE_GEPPA, A_PH_WA, unitcheck = TRUE)
}
\arguments{
\item{B_LU}{(character) The crop type}

\item{B_TEXTURE_GEPPA}{(character) The soil texture class in a particular region.}

\item{A_PH_WA}{(numeric) The pH measured in H2O}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The pH index in France estimated from pH in water, the textural class and the crop type
}
\description{
This function evaluates the pH index in France
}
