% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_nl}
\alias{osi_c_phosphor_nl}
\title{Calculate the phosphate availability index in the Netherlands}
\usage{
osi_c_phosphor_nl(
  B_LU,
  A_P_AL = NA_real_,
  A_P_CC = NA_real_,
  A_P_WA = NA_real_
)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with ammonium lactate (mg P/ kg)}

\item{A_P_CC}{(numeric) The P-content of the soil extracted with CaCl2 (mg P/ kg)}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water (mg P/ kg)}
}
\value{
The phosphate availability index in the Netherlands estimated from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate availability.
}
\examples{
osi_c_phosphor_nl(B_LU = 265, A_P_AL = 45, A_P_CC = 2.5)
osi_c_phosphor_nl(B_LU = c(265,1019),A_P_AL = c(35,54),A_P_CC = c(2.5,4.5), A_P_WA = c(35,65))

}
