% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_phosphor_no}
\alias{osi_c_phosphor_no}
\title{Calculate the phosphorus availability index in Norway}
\usage{
osi_c_phosphor_no(A_P_AL, B_LU = NA_character_)
}
\arguments{
\item{A_P_AL}{(numeric) The exchangeable P-content of the soil measured via Ammonium Lactate (mg P/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The phosphorus availability index in Norway estimated from extractable phosphorus. A numeric value.
}
\description{
This function calculates the phosphorus availability.
}
\examples{
osi_c_phosphor_no(B_LU = 'testcrop1', A_P_AL = 50)

}
