% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_phosphor.R
\name{osi_c_posphor}
\alias{osi_c_posphor}
\title{Calculate the phosphate availability index (wrapper function)}
\usage{
osi_c_posphor(
  B_LU,
  B_SOILTYPE_AGR = NA_character_,
  B_AER_FR = NA_character_,
  A_P_OL = NA_real_,
  A_P_AL = NA_real_,
  A_P_CC = NA_real_,
  A_P_WA = NA_real_,
  B_COUNTRY
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_AER_FR}{(character) An agroeconomic region in France}

\item{A_P_OL}{(numeric) The P-content of the soil extracted with Olsen (mg/kg)}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with ammonium lactate}

\item{A_P_CC}{(numeric) The P-content of the soil extracted with CaCl2}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water}

\item{B_COUNTRY}{(character) The country code}
}
\value{
The phosphate availability index in the Netherlands estimated from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate availability for all European countries (if available).
}
\examples{
osi_c_posphor(B_LU = 265, A_P_AL = 45, A_P_CC = 2.5,B_COUNTRY='NL')
osi_c_posphor(B_LU = 1019,A_P_AL = 54,A_P_CC = 4.5, A_P_WA = 65,B_COUNTRY ='NL')

}
