% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_fr}
\alias{osi_c_potassium_fr}
\title{Calculate the potassium availability index in France}
\usage{
osi_c_potassium_fr(
  B_LU,
  A_K_AAA,
  B_TEXTURE_GEPPA = NA_character_,
  B_SOILTYPE_AGR = NA_character_,
  B_AER_FR = NA_character_,
  A_PH_WA = NA_real_,
  unitcheck = TRUE
)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{A_K_AAA}{(numeric) The exchangeable K-content of the soil measured via ammonium acetate extraction}

\item{B_TEXTURE_GEPPA}{(character) The soil texture class in a particular region.}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_AER_FR}{(character) An agroeconomic region in France. Optional argument.}

\item{A_PH_WA}{(numeric) The pH measured in water.}

\item{unitcheck}{(character) Option to switch off unit checks (TRUE or FALSE)}
}
\value{
The potassium availability index in France estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\details{
The function has two optional arguments soil type (B_SOILTYPE_AGR) and agricultural region (B_AER_FR). When these are unknown, then the soil type is estimated based on the pH value. Threshold values are then generalized for calcareous and non-calcareous soils.
}
\examples{
osi_c_potassium_fr(B_LU = 'SOJ', A_K_AAA = 45,B_TEXTURE_GEPPA ='SS',
B_SOILTYPE_AGR = 'limons battants', B_AER_FR = 'nord picardie')

}
