% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_potassium.R
\name{osi_c_potassium_sk}
\alias{osi_c_potassium_sk}
\title{Calculate the potassium availability index in  Slovak Republic}
\usage{
osi_c_potassium_sk(B_TEXTURE_HYPRES, A_K_M3, B_LU = NA_character_)
}
\arguments{
\item{B_TEXTURE_HYPRES}{(character) The soil texture according to HYPRES classification system}

\item{A_K_M3}{(numeric) The exchangeable K-content of the soil measured via Mehlich 3 extracton (mg K/ kg)}

\item{B_LU}{(character) The crop code}
}
\value{
The potassium availability index in Slovak Republic estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium availability.
}
\examples{
osi_c_potassium_sk(A_K_M3 = 45,B_TEXTURE_HYPRES='C')

}
