% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_zinc.R
\name{osi_c_zinc_de}
\alias{osi_c_zinc_de}
\title{Calculate the zinc availability index in Germany}
\usage{
osi_c_zinc_de(B_LU, A_C_OF, A_CLAY_MI, A_SAND_MI, A_ZN_EDTA)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_C_OF}{(numeric) The carbon content of the soil layer (g/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_ZN_EDTA}{(numeric) Zn content measured in EDTA (mg / kg)}
}
\value{
The zinc availability index in Germany derived from extractable soil Zn fractions. A numeric value.
}
\description{
This function calculates the zinc availability.
}
\examples{
osi_c_zinc_de(B_LU = 265,A_C_OF=25, A_CLAY_MI=5,A_SAND_MI=15,A_ZN_EDTA = 50)
 
}
