% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_main.R
\name{osi_field}
\alias{osi_field}
\title{Calculate the Open Soil Index score for a single field}
\usage{
osi_field(
  B_LU,
  B_SOILTYPE_AGR,
  B_COUNTRY,
  A_SOM_LOI,
  A_SAND_MI,
  A_SILT_MI,
  A_CLAY_MI,
  A_PH_CC,
  A_N_RT,
  A_CN_FR,
  A_N_PMN,
  A_P_AL,
  A_P_CC,
  A_P_WA,
  ID = 1,
  output = "all"
)
}
\arguments{
\item{B_LU}{(numeric) a series with crop codes given the crop rotation plan (source: the BRP)}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_COUNTRY}{(character) The country code}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_CN_FR}{(numeric) The carbon to nitrogen ratio (-)}

\item{A_N_PMN}{(numeric) The potentially mineralizable N pool (mg N / kg soil)}

\item{A_P_AL}{(numeric) The P-AL content of the soil}

\item{A_P_CC}{(numeric) The plant available P content, extracted with 0.01M CaCl2 (mg / kg)}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)}

\item{ID}{(character) A field id}

\item{output}{(character) An optional argument to select output: obic_score, scores, indicators, recommendations, or all. (default = all)}
}
\value{
The output of the Open Soil Index Calculator for a specific agricultural field. 
Depending on the output type, different output objects can be returned.
These include the estimated OSI scores (both total and aggregated subscores), the value of the underling indicators as well the possible recommendations to improve the soil quality.
The output is always a data.table.
}
\description{
This functions wraps the functions of the OSI into one main function to calculate the score for Open Soil Index (OBI) for a single field.
}
\details{
It is assumed that the crop series is a continuous series in decreasing order of years. So most recent year first, oldest year last.
}
