% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_nut_potassium.R
\name{osi_nut_k_be}
\alias{osi_nut_k_be}
\title{Calculate the potassium excess index in Belgium}
\usage{
osi_nut_k_be(B_LU, B_TEXTURE_BE, A_K_AAA = NA_real_)
}
\arguments{
\item{B_LU}{(character) The crop code}

\item{B_TEXTURE_BE}{(character) The soil texture according to Belgium classification system}

\item{A_K_AAA}{(numeric) The exchangeable K-content of the soil measured via ammonium acetate extraction}
}
\value{
The potassium excess index in Belgium estimated from extractable potassium. A numeric value.
}
\description{
This function calculates the potassium excess.
}
\examples{
osi_nut_k_be(B_LU = 'SOJ',B_TEXTURE_BE, A_K_AAA = 45)

}
