% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_tables.R
\docType{data}
\name{osi_parms}
\alias{osi_parms}
\title{Table with OSI parameters being used in the package}
\format{
A data.table with x rows and x columns:
\describe{
  \item{osi_parm_id}{the parameter id}
  \item{osi_parm_name}{the name of the parameter}
  \item{osi_parm_type}{the type of the parameter. Options: measurement, field property}
  \item{osi_parm_description}{a short description of the parameters}
  \item{osi_parm_unit}{the unit of the parameter}
  \item{osi_parm_min}{the maximum allowed value for the parameter}
  \item{osi_parm_max}{the minimum allowed value for the parameter}
  \item{osi_parm_data_type}{the data type of the parameter: numeric, character or boolean}
  \item{osi_parm_enum}{does the parameter have predefined options}
  \item{osi_parm_options}{allowed options for the parameteer}
}
}
\usage{
osi_parms
}
\description{
This table contains all parameters being used in the OSI package to calculate the soil quality index.
}
\keyword{datasets}
